@@ -26,27 +26,29 @@
   },
   "homepage": "https://github.com/ProjectWyvern/wyvern-js#readme",
   "dependencies": {
-    "0x.js": "^0.29.2",
-    "@0xproject/assert": "^0.0.11",
-    "@0xproject/json-schemas": "^0.7.3",
-    "@0xproject/utils": "^0.1.3",
-    "@0xproject/web3-wrapper": "^0.1.4",
+    "@0x/assert": "^3.0.30",
+    "@0x/base-contract": "^6.4.4",
+    "@0x/json-schemas": "^6.4.0",
+    "@0x/utils": "https://github.com/ProjectOpensea/0x-tools/raw/provider-patch/utils/0x-utils-6.5.0.tgz",
+    "@0x/web3-wrapper": "^7.6.1",
     "bn.js": "^4.11.8",
     "ethereumjs-abi": "https://github.com/ProjectWyvern/ethereumjs-abi.git",
     "ethereumjs-util": "^5.1.3",
+    "ethers": "^4.0.49",
     "json-loader": "^0.5.7",
     "jsonschema": "^1.2.2",
-    "lodash": "^4.17.4",
-    "web3": "^0.20.4"
+    "lodash": "^4.17.21",
+    "web3": "^1.7.0"
   },
   "devDependencies": {
-    "@0xproject/abi-gen": "0.1.2",
-    "@0xproject/types": "^0.1.3",
+    "@0x/abi-gen": "^5.7.0",
+    "@0x/types": "^3.3.4",
     "@types/chai": "^4.1.3",
-    "@types/lodash": "^4.14.92",
+    "@types/lodash": "^4.14.178",
     "@types/node": "16.11.0",
     "awesome-typescript-loader": "^3.4.1",
     "chai": "^4.1.2",
+    "ethereum-types": "^3.6.0",
     "mocha": "^5.2.0",
     "mocha-typescript": "^1.1.15",
     "ts-node": "^6.1.0",
@@ -55,9 +57,8 @@
     "tslint-react": "^5.0.0",
     "types-bn": "^0.0.1",
     "types-ethereumjs-util": "^0.0.5",
-    "typescript": "^2.6.2",
+    "typescript": "^4.5.4",
     "typescript-eslint-parser": "^16.0.0",
-    "web3-typescript-typings": "^0.9.3",
     "webpack": "^3.10.0"
   }
 }
