@@ -1,12 +1,13 @@
 "use strict";
 /* Sourced from 0x.js */
 Object.defineProperty(exports, "__esModule", { value: true });
-var utils_1 = require("@0xproject/utils");
-var BN = require("bn.js");
-var ethABI = require("ethereumjs-abi");
-var ethUtil = require("ethereumjs-util");
-var _ = require("lodash");
-var types_1 = require("../types");
+exports.utils = void 0;
+const utils_1 = require("@0x/utils");
+const BN = require("bn.js");
+const ethABI = require("ethereumjs-abi");
+const ethUtil = require("ethereumjs-util");
+const _ = require("lodash");
+const types_1 = require("../types");
 exports.utils = {
     /**
      * Converts BigNumber instance to BN
@@ -14,29 +15,29 @@ exports.utils = {
      * expects values of Solidity type `uint` to be passed as type `BN`.
      * We do not use BN anywhere else in the codebase.
      */
-    bigNumberToBN: function (value) {
+    bigNumberToBN(value) {
         return new BN(value.toString(), 10);
     },
-    consoleLog: function (message) {
+    consoleLog(message) {
         // tslint:disable-next-line: no-console
         console.log(message);
     },
-    isParityNode: function (nodeVersion) {
+    isParityNode(nodeVersion) {
         return _.includes(nodeVersion, 'Parity');
     },
-    isTestRpc: function (nodeVersion) {
+    isTestRpc(nodeVersion) {
         return _.includes(nodeVersion, 'TestRPC');
     },
-    spawnSwitchErr: function (name, value) {
-        return new Error("Unexpected switch value: " + value + " encountered for " + name);
+    spawnSwitchErr(name, value) {
+        return new Error(`Unexpected switch value: ${value} encountered for ${name}`);
     },
-    getAssetHashHex: function (assetHash, schema) {
-        var concat = schema + ':' + assetHash;
-        var hashBuf = ethABI.soliditySHA3(['string'], [concat]);
+    getAssetHashHex(assetHash, schema) {
+        const concat = schema + ':' + assetHash;
+        const hashBuf = ethABI.soliditySHA3(['string'], [concat]);
         return ethUtil.bufferToHex(hashBuf);
     },
-    getOrderHashHex: function (order) {
-        var orderParts = [
+    getOrderHashHex(order) {
+        const orderParts = [
             { value: order.exchange, type: types_1.SolidityTypes.Address },
             { value: order.maker, type: types_1.SolidityTypes.Address },
             { value: order.taker, type: types_1.SolidityTypes.Address },
@@ -50,10 +51,10 @@ exports.utils = {
             { value: order.saleKind, type: types_1.SolidityTypes.Uint8 },
             { value: order.target, type: types_1.SolidityTypes.Address },
             { value: order.howToCall, type: types_1.SolidityTypes.Uint8 },
-            { value: new Buffer(order.calldata.slice(2), 'hex'), type: types_1.SolidityTypes.Bytes },
-            { value: new Buffer(order.replacementPattern.slice(2), 'hex'), type: types_1.SolidityTypes.Bytes },
+            { value: Buffer.from(order.calldata.slice(2), 'hex'), type: types_1.SolidityTypes.Bytes },
+            { value: Buffer.from(order.replacementPattern.slice(2), 'hex'), type: types_1.SolidityTypes.Bytes },
             { value: order.staticTarget, type: types_1.SolidityTypes.Address },
-            { value: new Buffer(order.staticExtradata.slice(2), 'hex'), type: types_1.SolidityTypes.Bytes },
+            { value: Buffer.from(order.staticExtradata.slice(2), 'hex'), type: types_1.SolidityTypes.Bytes },
             { value: order.paymentToken, type: types_1.SolidityTypes.Address },
             { value: exports.utils.bigNumberToBN(order.basePrice), type: types_1.SolidityTypes.Uint256 },
             { value: exports.utils.bigNumberToBN(order.extra), type: types_1.SolidityTypes.Uint256 },
@@ -61,15 +62,15 @@ exports.utils = {
             { value: exports.utils.bigNumberToBN(order.expirationTime), type: types_1.SolidityTypes.Uint256 },
             { value: exports.utils.bigNumberToBN(order.salt), type: types_1.SolidityTypes.Uint256 },
         ];
-        var types = _.map(orderParts, function (o) { return o.type; });
-        var values = _.map(orderParts, function (o) { return o.value; });
-        var hash = ethABI.soliditySHA3(types, values);
+        const types = _.map(orderParts, o => o.type);
+        const values = _.map(orderParts, o => o.value);
+        const hash = ethABI.soliditySHA3(types, values);
         return ethUtil.bufferToHex(hash);
     },
-    getCurrentUnixTimestampSec: function () {
-        return new utils_1.BigNumber(Date.now() / 1000).round();
+    getCurrentUnixTimestampSec() {
+        return new utils_1.BigNumber(Date.now() / 1000).integerValue();
     },
-    getCurrentUnixTimestampMs: function () {
+    getCurrentUnixTimestampMs() {
         return new utils_1.BigNumber(Date.now());
     },
 };
