@@ -1,393 +1,849 @@
 "use strict";
-var __extends = (this && this.__extends) || (function () {
-    var extendStatics = Object.setPrototypeOf ||
-        ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-        function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
-    return function (d, b) {
-        extendStatics(d, b);
-        function __() { this.constructor = d; }
-        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-    };
-})();
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (_) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
 Object.defineProperty(exports, "__esModule", { value: true });
-var utils_1 = require("@0xproject/utils");
-var base_contract_1 = require("./base_contract");
-var AuthenticatedProxyContract = /** @class */ (function (_super) {
-    __extends(AuthenticatedProxyContract, _super);
-    function AuthenticatedProxyContract(web3ContractInstance, defaults) {
-        var _this = _super.call(this, web3ContractInstance, defaults) || this;
-        _this.proxyType = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.proxyType.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.user = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.user.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.implementation = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.implementation.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.revoked = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.revoked.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.upgradeabilityOwner = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.upgradeabilityOwner.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.registry = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.registry.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.receiveApproval = {
-            sendTransactionAsync: function (from_0, value_1, token_2, extraData_3, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.receiveApproval.estimateGasAsync.bind(self, from_0, value_1, token_2, extraData_3))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.receiveApproval, self.web3ContractInstance)(from_0, value_1, token_2, extraData_3, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (from_0, value_1, token_2, extraData_3, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.receiveApproval.estimateGas, self.web3ContractInstance)(from_0, value_1, token_2, extraData_3, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (from_0, value_1, token_2, extraData_3) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.receiveApproval.getData(from_0, value_1, token_2, extraData_3);
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.initialize = {
-            sendTransactionAsync: function (addrUser_0, addrRegistry_1, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.initialize.estimateGasAsync.bind(self, addrUser_0, addrRegistry_1))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.initialize, self.web3ContractInstance)(addrUser_0, addrRegistry_1, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (addrUser_0, addrRegistry_1, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.initialize.estimateGas, self.web3ContractInstance)(addrUser_0, addrRegistry_1, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (addrUser_0, addrRegistry_1) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.initialize.getData(addrUser_0, addrRegistry_1);
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.setRevoke = {
-            sendTransactionAsync: function (revoke_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.setRevoke.estimateGasAsync.bind(self, revoke_0))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.setRevoke, self.web3ContractInstance)(revoke_0, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (revoke_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.setRevoke.estimateGas, self.web3ContractInstance)(revoke_0, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (revoke_0) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.setRevoke.getData(revoke_0);
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.proxy = {
-            sendTransactionAsync: function (dest_0, howToCall_1, calldata_2, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.proxy.estimateGasAsync.bind(self, dest_0, howToCall_1, calldata_2))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.proxy, self.web3ContractInstance)(dest_0, howToCall_1, calldata_2, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (dest_0, howToCall_1, calldata_2, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.proxy.estimateGas, self.web3ContractInstance)(dest_0, howToCall_1, calldata_2, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (dest_0, howToCall_1, calldata_2) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.proxy.getData(dest_0, howToCall_1, calldata_2);
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.proxyAssert = {
-            sendTransactionAsync: function (dest_0, howToCall_1, calldata_2, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.proxyAssert.estimateGasAsync.bind(self, dest_0, howToCall_1, calldata_2))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.proxyAssert, self.web3ContractInstance)(dest_0, howToCall_1, calldata_2, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (dest_0, howToCall_1, calldata_2, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.proxyAssert.estimateGas, self.web3ContractInstance)(dest_0, howToCall_1, calldata_2, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (dest_0, howToCall_1, calldata_2) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.proxyAssert.getData(dest_0, howToCall_1, calldata_2);
-                return abiEncodedTransactionData;
-            },
-        };
-        utils_1.classUtils.bindAll(_this, ['web3ContractInstance', 'defaults']);
-        return _this;
+exports.AuthenticatedProxyContract = exports.AuthenticatedProxyEvents = void 0;
+// tslint:disable:no-consecutive-blank-lines ordered-imports align trailing-comma enum-naming
+// tslint:disable:whitespace no-unbound-method no-trailing-whitespace
+// tslint:disable:no-unused-variable
+const base_contract_1 = require("@0x/base-contract");
+const json_schemas_1 = require("@0x/json-schemas");
+const utils_1 = require("@0x/utils");
+const web3_wrapper_1 = require("@0x/web3-wrapper");
+const assert_1 = require("@0x/assert");
+const ethers = require("ethers");
+var AuthenticatedProxyEvents;
+(function (AuthenticatedProxyEvents) {
+    AuthenticatedProxyEvents["Revoked"] = "Revoked";
+    AuthenticatedProxyEvents["ReceivedEther"] = "ReceivedEther";
+    AuthenticatedProxyEvents["ReceivedTokens"] = "ReceivedTokens";
+})(AuthenticatedProxyEvents = exports.AuthenticatedProxyEvents || (exports.AuthenticatedProxyEvents = {}));
+/* istanbul ignore next */
+// tslint:disable:array-type
+// tslint:disable:no-parameter-reassignment
+// tslint:disable-next-line:class-name
+class AuthenticatedProxyContract extends base_contract_1.BaseContract {
+    constructor(address, supportedProvider, txDefaults, logDecodeDependencies, deployedBytecode = AuthenticatedProxyContract.deployedBytecode, encoderOverrides) {
+        super('AuthenticatedProxy', AuthenticatedProxyContract.ABI(), address, supportedProvider, txDefaults, logDecodeDependencies, deployedBytecode, encoderOverrides);
+        this._methodABIIndex = {};
+        utils_1.classUtils.bindAll(this, ['_abiEncoderByFunctionSignature', 'address', '_web3Wrapper']);
+        this._subscriptionManager = new base_contract_1.SubscriptionManager(AuthenticatedProxyContract.ABI(), this._web3Wrapper);
+        AuthenticatedProxyContract.ABI().forEach((item, index) => {
+            if (item.type === 'function') {
+                const methodAbi = item;
+                this._methodABIIndex[methodAbi.name] = index;
+            }
+        });
     }
-    return AuthenticatedProxyContract;
-}(base_contract_1.BaseContract)); // tslint:disable:max-file-line-count
+    static deployFrom0xArtifactAsync(artifact, supportedProvider, txDefaults, logDecodeDependencies) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.doesConformToSchema('txDefaults', txDefaults, json_schemas_1.schemas.txDataSchema);
+            if (artifact.compilerOutput === undefined) {
+                throw new Error('Compiler output not found in the artifact file');
+            }
+            const provider = utils_1.providerUtils.standardizeOrThrow(supportedProvider);
+            const bytecode = artifact.compilerOutput.evm.bytecode.object;
+            const abi = artifact.compilerOutput.abi;
+            const logDecodeDependenciesAbiOnly = {};
+            if (Object.keys(logDecodeDependencies) !== undefined) {
+                for (const key of Object.keys(logDecodeDependencies)) {
+                    logDecodeDependenciesAbiOnly[key] = logDecodeDependencies[key].compilerOutput.abi;
+                }
+            }
+            return AuthenticatedProxyContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly);
+        });
+    }
+    static deployWithLibrariesFrom0xArtifactAsync(artifact, libraryArtifacts, supportedProvider, txDefaults, logDecodeDependencies) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.doesConformToSchema('txDefaults', txDefaults, json_schemas_1.schemas.txDataSchema);
+            if (artifact.compilerOutput === undefined) {
+                throw new Error('Compiler output not found in the artifact file');
+            }
+            const provider = utils_1.providerUtils.standardizeOrThrow(supportedProvider);
+            const abi = artifact.compilerOutput.abi;
+            const logDecodeDependenciesAbiOnly = {};
+            if (Object.keys(logDecodeDependencies) !== undefined) {
+                for (const key of Object.keys(logDecodeDependencies)) {
+                    logDecodeDependenciesAbiOnly[key] = logDecodeDependencies[key].compilerOutput.abi;
+                }
+            }
+            const libraryAddresses = yield AuthenticatedProxyContract._deployLibrariesAsync(artifact, libraryArtifacts, new web3_wrapper_1.Web3Wrapper(provider), txDefaults);
+            const bytecode = (0, base_contract_1.linkLibrariesInBytecode)(artifact, libraryAddresses);
+            return AuthenticatedProxyContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly);
+        });
+    }
+    static deployAsync(bytecode, abi, supportedProvider, txDefaults, logDecodeDependencies) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.isHexString('bytecode', bytecode);
+            assert_1.assert.doesConformToSchema('txDefaults', txDefaults, json_schemas_1.schemas.txDataSchema);
+            const provider = utils_1.providerUtils.standardizeOrThrow(supportedProvider);
+            const constructorAbi = base_contract_1.BaseContract._lookupConstructorAbi(abi);
+            [] = base_contract_1.BaseContract._formatABIDataItemList(constructorAbi.inputs, [], base_contract_1.BaseContract._bigNumberToString);
+            const iface = new ethers.utils.Interface(abi);
+            const deployInfo = iface.deployFunction;
+            const txData = deployInfo.encode(bytecode, []);
+            const web3Wrapper = new web3_wrapper_1.Web3Wrapper(provider);
+            const txDataWithDefaults = yield base_contract_1.BaseContract._applyDefaultsToContractTxDataAsync(Object.assign({ data: txData }, txDefaults), web3Wrapper.estimateGasAsync.bind(web3Wrapper));
+            const txHash = yield web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+            utils_1.logUtils.log(`transactionHash: ${txHash}`);
+            const txReceipt = yield web3Wrapper.awaitTransactionSuccessAsync(txHash);
+            utils_1.logUtils.log(`AuthenticatedProxy successfully deployed at ${txReceipt.contractAddress}`);
+            const contractInstance = new AuthenticatedProxyContract(txReceipt.contractAddress, provider, txDefaults, logDecodeDependencies);
+            contractInstance.constructorArgs = [];
+            return contractInstance;
+        });
+    }
+    /**
+     * @returns      The contract ABI
+     */
+    static ABI() {
+        const abi = [
+            {
+                constant: true,
+                inputs: [],
+                name: 'proxyType',
+                outputs: [
+                    {
+                        name: 'proxyTypeId',
+                        type: 'uint256',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'pure',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [],
+                name: 'user',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [],
+                name: 'implementation',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [],
+                name: 'revoked',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'bool',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [],
+                name: 'upgradeabilityOwner',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [],
+                name: 'registry',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'from',
+                        type: 'address',
+                    },
+                    {
+                        name: 'value',
+                        type: 'uint256',
+                    },
+                    {
+                        name: 'token',
+                        type: 'address',
+                    },
+                    {
+                        name: 'extraData',
+                        type: 'bytes',
+                    },
+                ],
+                name: 'receiveApproval',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                inputs: [],
+                outputs: [],
+                payable: true,
+                stateMutability: 'payable',
+                type: 'fallback',
+            },
+            {
+                anonymous: false,
+                inputs: [
+                    {
+                        name: 'revoked',
+                        type: 'bool',
+                        indexed: false,
+                    },
+                ],
+                name: 'Revoked',
+                outputs: [],
+                type: 'event',
+            },
+            {
+                anonymous: false,
+                inputs: [
+                    {
+                        name: 'sender',
+                        type: 'address',
+                        indexed: true,
+                    },
+                    {
+                        name: 'amount',
+                        type: 'uint256',
+                        indexed: false,
+                    },
+                ],
+                name: 'ReceivedEther',
+                outputs: [],
+                type: 'event',
+            },
+            {
+                anonymous: false,
+                inputs: [
+                    {
+                        name: 'from',
+                        type: 'address',
+                        indexed: true,
+                    },
+                    {
+                        name: 'value',
+                        type: 'uint256',
+                        indexed: false,
+                    },
+                    {
+                        name: 'token',
+                        type: 'address',
+                        indexed: true,
+                    },
+                    {
+                        name: 'extraData',
+                        type: 'bytes',
+                        indexed: false,
+                    },
+                ],
+                name: 'ReceivedTokens',
+                outputs: [],
+                type: 'event',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'addrUser',
+                        type: 'address',
+                    },
+                    {
+                        name: 'addrRegistry',
+                        type: 'address',
+                    },
+                ],
+                name: 'initialize',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'revoke',
+                        type: 'bool',
+                    },
+                ],
+                name: 'setRevoke',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'dest',
+                        type: 'address',
+                    },
+                    {
+                        name: 'howToCall',
+                        type: 'uint8',
+                    },
+                    {
+                        name: 'calldata',
+                        type: 'bytes',
+                    },
+                ],
+                name: 'proxy',
+                outputs: [
+                    {
+                        name: 'result',
+                        type: 'bool',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'dest',
+                        type: 'address',
+                    },
+                    {
+                        name: 'howToCall',
+                        type: 'uint8',
+                    },
+                    {
+                        name: 'calldata',
+                        type: 'bytes',
+                    },
+                ],
+                name: 'proxyAssert',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+        ];
+        return abi;
+    }
+    static _deployLibrariesAsync(artifact, libraryArtifacts, web3Wrapper, txDefaults, libraryAddresses = {}) {
+        return __awaiter(this, void 0, void 0, function* () {
+            const links = artifact.compilerOutput.evm.bytecode.linkReferences;
+            // Go through all linked libraries, recursively deploying them if necessary.
+            for (const link of Object.values(links)) {
+                for (const libraryName of Object.keys(link)) {
+                    if (!libraryAddresses[libraryName]) {
+                        // Library not yet deployed.
+                        const libraryArtifact = libraryArtifacts[libraryName];
+                        if (!libraryArtifact) {
+                            throw new Error(`Missing artifact for linked library "${libraryName}"`);
+                        }
+                        // Deploy any dependent libraries used by this library.
+                        yield AuthenticatedProxyContract._deployLibrariesAsync(libraryArtifact, libraryArtifacts, web3Wrapper, txDefaults, libraryAddresses);
+                        // Deploy this library.
+                        const linkedLibraryBytecode = (0, base_contract_1.linkLibrariesInBytecode)(libraryArtifact, libraryAddresses);
+                        const txDataWithDefaults = yield base_contract_1.BaseContract._applyDefaultsToContractTxDataAsync(Object.assign({ data: linkedLibraryBytecode }, txDefaults), web3Wrapper.estimateGasAsync.bind(web3Wrapper));
+                        const txHash = yield web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                        utils_1.logUtils.log(`transactionHash: ${txHash}`);
+                        const { contractAddress } = yield web3Wrapper.awaitTransactionSuccessAsync(txHash);
+                        utils_1.logUtils.log(`${libraryArtifact.contractName} successfully deployed at ${contractAddress}`);
+                        libraryAddresses[libraryArtifact.contractName] = contractAddress;
+                    }
+                }
+            }
+            return libraryAddresses;
+        });
+    }
+    getFunctionSignature(methodName) {
+        const index = this._methodABIIndex[methodName];
+        const methodAbi = AuthenticatedProxyContract.ABI()[index]; // tslint:disable-line:no-unnecessary-type-assertion
+        const functionSignature = (0, base_contract_1.methodAbiToFunctionSignature)(methodAbi);
+        return functionSignature;
+    }
+    getABIDecodedTransactionData(methodName, callData) {
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = this;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        const abiDecodedCallData = abiEncoder.strictDecode(callData);
+        return abiDecodedCallData;
+    }
+    getABIDecodedReturnData(methodName, callData) {
+        if (this._encoderOverrides.decodeOutput) {
+            return this._encoderOverrides.decodeOutput(methodName, callData);
+        }
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = this;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        const abiDecodedCallData = abiEncoder.strictDecodeReturnValue(callData);
+        return abiDecodedCallData;
+    }
+    getSelector(methodName) {
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = this;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        return abiEncoder.getSelector();
+    }
+    proxyType() {
+        const self = this;
+        const functionSignature = 'proxyType()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    let rawCallResult;
+                    if (self._deployedBytecodeIfExists) {
+                        rawCallResult = yield self._evmExecAsync(this.getABIEncodedTransactionData());
+                    }
+                    else {
+                        rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    }
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    user() {
+        const self = this;
+        const functionSignature = 'user()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    implementation() {
+        const self = this;
+        const functionSignature = 'implementation()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    revoked() {
+        const self = this;
+        const functionSignature = 'revoked()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    upgradeabilityOwner() {
+        const self = this;
+        const functionSignature = 'upgradeabilityOwner()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    registry() {
+        const self = this;
+        const functionSignature = 'registry()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    receiveApproval(from, value, token, extraData) {
+        const self = this;
+        assert_1.assert.isString('from', from);
+        assert_1.assert.isBigNumber('value', value);
+        assert_1.assert.isString('token', token);
+        assert_1.assert.isString('extraData', extraData);
+        const functionSignature = 'receiveApproval(address,uint256,address,bytes)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [from.toLowerCase(),
+                    value,
+                    token.toLowerCase(),
+                    extraData
+                ]);
+            },
+        };
+    }
+    initialize(addrUser, addrRegistry) {
+        const self = this;
+        assert_1.assert.isString('addrUser', addrUser);
+        assert_1.assert.isString('addrRegistry', addrRegistry);
+        const functionSignature = 'initialize(address,address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [addrUser.toLowerCase(),
+                    addrRegistry.toLowerCase()
+                ]);
+            },
+        };
+    }
+    setRevoke(revoke) {
+        const self = this;
+        assert_1.assert.isBoolean('revoke', revoke);
+        const functionSignature = 'setRevoke(bool)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [revoke
+                ]);
+            },
+        };
+    }
+    proxy(dest, howToCall, calldata) {
+        const self = this;
+        assert_1.assert.isString('dest', dest);
+        assert_1.assert.isNumberOrBigNumber('howToCall', howToCall);
+        assert_1.assert.isString('calldata', calldata);
+        const functionSignature = 'proxy(address,uint8,bytes)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [dest.toLowerCase(),
+                    howToCall,
+                    calldata
+                ]);
+            },
+        };
+    }
+    proxyAssert(dest, howToCall, calldata) {
+        const self = this;
+        assert_1.assert.isString('dest', dest);
+        assert_1.assert.isNumberOrBigNumber('howToCall', howToCall);
+        assert_1.assert.isString('calldata', calldata);
+        const functionSignature = 'proxyAssert(address,uint8,bytes)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [dest.toLowerCase(),
+                    howToCall,
+                    calldata
+                ]);
+            },
+        };
+    }
+    /**
+     * Subscribe to an event type emitted by the AuthenticatedProxy contract.
+     * @param eventName The AuthenticatedProxy contract event you would like to subscribe to.
+     * @param indexFilterValues An object where the keys are indexed args returned by the event and
+     * the value is the value you are interested in. E.g `{maker: aUserAddressHex}`
+     * @param callback Callback that gets called when a log is added/removed
+     * @param isVerbose Enable verbose subscription warnings (e.g recoverable network issues encountered)
+     * @return Subscription token used later to unsubscribe
+     */
+    subscribe(eventName, indexFilterValues, callback, isVerbose = false, blockPollingIntervalMs) {
+        assert_1.assert.doesBelongToStringEnum('eventName', eventName, AuthenticatedProxyEvents);
+        assert_1.assert.doesConformToSchema('indexFilterValues', indexFilterValues, json_schemas_1.schemas.indexFilterValuesSchema);
+        assert_1.assert.isFunction('callback', callback);
+        const subscriptionToken = this._subscriptionManager.subscribe(this.address, eventName, indexFilterValues, AuthenticatedProxyContract.ABI(), callback, isVerbose, blockPollingIntervalMs);
+        return subscriptionToken;
+    }
+    /**
+     * Cancel a subscription
+     * @param subscriptionToken Subscription token returned by `subscribe()`
+     */
+    unsubscribe(subscriptionToken) {
+        this._subscriptionManager.unsubscribe(subscriptionToken);
+    }
+    /**
+     * Cancels all existing subscriptions
+     */
+    unsubscribeAll() {
+        this._subscriptionManager.unsubscribeAll();
+    }
+    /**
+     * Gets historical logs without creating a subscription
+     * @param eventName The AuthenticatedProxy contract event you would like to subscribe to.
+     * @param blockRange Block range to get logs from.
+     * @param indexFilterValues An object where the keys are indexed args returned by the event and
+     * the value is the value you are interested in. E.g `{_from: aUserAddressHex}`
+     * @return Array of logs that match the parameters
+     */
+    getLogsAsync(eventName, blockRange, indexFilterValues) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.doesBelongToStringEnum('eventName', eventName, AuthenticatedProxyEvents);
+            assert_1.assert.doesConformToSchema('blockRange', blockRange, json_schemas_1.schemas.blockRangeSchema);
+            assert_1.assert.doesConformToSchema('indexFilterValues', indexFilterValues, json_schemas_1.schemas.indexFilterValuesSchema);
+            const logs = yield this._subscriptionManager.getLogsAsync(this.address, eventName, blockRange, indexFilterValues, AuthenticatedProxyContract.ABI());
+            return logs;
+        });
+    }
+}
 exports.AuthenticatedProxyContract = AuthenticatedProxyContract;
+AuthenticatedProxyContract.contractName = 'AuthenticatedProxy';
+// tslint:disable:max-file-line-count
+// tslint:enable:no-unbound-method no-parameter-reassignment no-consecutive-blank-lines ordered-imports align
+// tslint:enable:trailing-comma whitespace no-trailing-whitespace
 //# sourceMappingURL=authenticated_proxy.js.map
