@@ -1,16 +1,43 @@
-/**
- * This file is auto-generated using abi-gen. Don't edit directly.
- * Templates can be found at https://github.com/0xProject/0x.js/tree/development/packages/abi-gen-templates.
- */
-import { TxData } from '@0xproject/types';
-import { BigNumber } from '@0xproject/utils';
-import * as Web3 from 'web3';
-import { BaseContract } from './base_contract';
+import { EncoderOverrides, ContractTxFunctionObj, BaseContract } from '@0x/base-contract';
+import { ContractAbi, ContractArtifact, TxData, SupportedProvider } from 'ethereum-types';
+import { BigNumber } from '@0x/utils';
+import { SimpleContractArtifact } from '@0x/types';
+import { Web3Wrapper } from '@0x/web3-wrapper';
 export declare class WyvernAtomicizerContract extends BaseContract {
-    atomicize: {
-        sendTransactionAsync(addrs_0: string[], values_1: BigNumber[], calldataLengths_2: BigNumber[], calldatas_3: string, txData?: TxData): Promise<string>;
-        estimateGasAsync(addrs_0: string[], values_1: BigNumber[], calldataLengths_2: BigNumber[], calldatas_3: string, txData?: TxData): Promise<number>;
-        getABIEncodedTransactionData(addrs_0: string[], values_1: BigNumber[], calldataLengths_2: BigNumber[], calldatas_3: string): string;
-    };
-    constructor(web3ContractInstance: Web3.ContractInstance, defaults: Partial<TxData>);
+    /**
+     * @ignore
+     */
+    static deployedBytecode: string | undefined;
+    static contractName: string;
+    private readonly _methodABIIndex;
+    static deployFrom0xArtifactAsync(artifact: ContractArtifact | SimpleContractArtifact, supportedProvider: SupportedProvider, txDefaults: Partial<TxData>, logDecodeDependencies: {
+        [contractName: string]: (ContractArtifact | SimpleContractArtifact);
+    }): Promise<WyvernAtomicizerContract>;
+    static deployWithLibrariesFrom0xArtifactAsync(artifact: ContractArtifact, libraryArtifacts: {
+        [libraryName: string]: ContractArtifact;
+    }, supportedProvider: SupportedProvider, txDefaults: Partial<TxData>, logDecodeDependencies: {
+        [contractName: string]: (ContractArtifact | SimpleContractArtifact);
+    }): Promise<WyvernAtomicizerContract>;
+    static deployAsync(bytecode: string, abi: ContractAbi, supportedProvider: SupportedProvider, txDefaults: Partial<TxData>, logDecodeDependencies: {
+        [contractName: string]: ContractAbi;
+    }): Promise<WyvernAtomicizerContract>;
+    /**
+     * @returns      The contract ABI
+     */
+    static ABI(): ContractAbi;
+    protected static _deployLibrariesAsync(artifact: ContractArtifact, libraryArtifacts: {
+        [libraryName: string]: ContractArtifact;
+    }, web3Wrapper: Web3Wrapper, txDefaults: Partial<TxData>, libraryAddresses?: {
+        [libraryName: string]: string;
+    }): Promise<{
+        [libraryName: string]: string;
+    }>;
+    getFunctionSignature(methodName: string): string;
+    getABIDecodedTransactionData<T>(methodName: string, callData: string): T;
+    getABIDecodedReturnData<T>(methodName: string, callData: string): T;
+    getSelector(methodName: string): string;
+    atomicize(addrs: string[], values: BigNumber[], calldataLengths: BigNumber[], calldatas: string): ContractTxFunctionObj<void>;
+    constructor(address: string, supportedProvider: SupportedProvider, txDefaults?: Partial<TxData>, logDecodeDependencies?: {
+        [contractName: string]: ContractAbi;
+    }, deployedBytecode?: string | undefined, encoderOverrides?: Partial<EncoderOverrides>);
 }
