@@ -1,105 +1,246 @@
 "use strict";
-var __extends = (this && this.__extends) || (function () {
-    var extendStatics = Object.setPrototypeOf ||
-        ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-        function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
-    return function (d, b) {
-        extendStatics(d, b);
-        function __() { this.constructor = d; }
-        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-    };
-})();
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (_) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
 Object.defineProperty(exports, "__esModule", { value: true });
-var utils_1 = require("@0xproject/utils");
-var base_contract_1 = require("./base_contract");
-var WyvernAtomicizerContract = /** @class */ (function (_super) {
-    __extends(WyvernAtomicizerContract, _super);
-    function WyvernAtomicizerContract(web3ContractInstance, defaults) {
-        var _this = _super.call(this, web3ContractInstance, defaults) || this;
-        _this.atomicize = {
-            sendTransactionAsync: function (addrs_0, values_1, calldataLengths_2, calldatas_3, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.atomicize.estimateGasAsync.bind(self, addrs_0, values_1, calldataLengths_2, calldatas_3))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.atomicize, self.web3ContractInstance)(addrs_0, values_1, calldataLengths_2, calldatas_3, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
+exports.WyvernAtomicizerContract = void 0;
+// tslint:disable:no-consecutive-blank-lines ordered-imports align trailing-comma enum-naming
+// tslint:disable:whitespace no-unbound-method no-trailing-whitespace
+// tslint:disable:no-unused-variable
+const base_contract_1 = require("@0x/base-contract");
+const json_schemas_1 = require("@0x/json-schemas");
+const utils_1 = require("@0x/utils");
+const web3_wrapper_1 = require("@0x/web3-wrapper");
+const assert_1 = require("@0x/assert");
+const ethers = require("ethers");
+// tslint:enable:no-unused-variable
+/* istanbul ignore next */
+// tslint:disable:array-type
+// tslint:disable:no-parameter-reassignment
+// tslint:disable-next-line:class-name
+class WyvernAtomicizerContract extends base_contract_1.BaseContract {
+    constructor(address, supportedProvider, txDefaults, logDecodeDependencies, deployedBytecode = WyvernAtomicizerContract.deployedBytecode, encoderOverrides) {
+        super('WyvernAtomicizer', WyvernAtomicizerContract.ABI(), address, supportedProvider, txDefaults, logDecodeDependencies, deployedBytecode, encoderOverrides);
+        this._methodABIIndex = {};
+        utils_1.classUtils.bindAll(this, ['_abiEncoderByFunctionSignature', 'address', '_web3Wrapper']);
+        WyvernAtomicizerContract.ABI().forEach((item, index) => {
+            if (item.type === 'function') {
+                const methodAbi = item;
+                this._methodABIIndex[methodAbi.name] = index;
+            }
+        });
+    }
+    static deployFrom0xArtifactAsync(artifact, supportedProvider, txDefaults, logDecodeDependencies) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.doesConformToSchema('txDefaults', txDefaults, json_schemas_1.schemas.txDataSchema);
+            if (artifact.compilerOutput === undefined) {
+                throw new Error('Compiler output not found in the artifact file');
+            }
+            const provider = utils_1.providerUtils.standardizeOrThrow(supportedProvider);
+            const bytecode = artifact.compilerOutput.evm.bytecode.object;
+            const abi = artifact.compilerOutput.abi;
+            const logDecodeDependenciesAbiOnly = {};
+            if (Object.keys(logDecodeDependencies) !== undefined) {
+                for (const key of Object.keys(logDecodeDependencies)) {
+                    logDecodeDependenciesAbiOnly[key] = logDecodeDependencies[key].compilerOutput.abi;
+                }
+            }
+            return WyvernAtomicizerContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly);
+        });
+    }
+    static deployWithLibrariesFrom0xArtifactAsync(artifact, libraryArtifacts, supportedProvider, txDefaults, logDecodeDependencies) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.doesConformToSchema('txDefaults', txDefaults, json_schemas_1.schemas.txDataSchema);
+            if (artifact.compilerOutput === undefined) {
+                throw new Error('Compiler output not found in the artifact file');
+            }
+            const provider = utils_1.providerUtils.standardizeOrThrow(supportedProvider);
+            const abi = artifact.compilerOutput.abi;
+            const logDecodeDependenciesAbiOnly = {};
+            if (Object.keys(logDecodeDependencies) !== undefined) {
+                for (const key of Object.keys(logDecodeDependencies)) {
+                    logDecodeDependenciesAbiOnly[key] = logDecodeDependencies[key].compilerOutput.abi;
+                }
+            }
+            const libraryAddresses = yield WyvernAtomicizerContract._deployLibrariesAsync(artifact, libraryArtifacts, new web3_wrapper_1.Web3Wrapper(provider), txDefaults);
+            const bytecode = (0, base_contract_1.linkLibrariesInBytecode)(artifact, libraryAddresses);
+            return WyvernAtomicizerContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly);
+        });
+    }
+    static deployAsync(bytecode, abi, supportedProvider, txDefaults, logDecodeDependencies) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.isHexString('bytecode', bytecode);
+            assert_1.assert.doesConformToSchema('txDefaults', txDefaults, json_schemas_1.schemas.txDataSchema);
+            const provider = utils_1.providerUtils.standardizeOrThrow(supportedProvider);
+            const constructorAbi = base_contract_1.BaseContract._lookupConstructorAbi(abi);
+            [] = base_contract_1.BaseContract._formatABIDataItemList(constructorAbi.inputs, [], base_contract_1.BaseContract._bigNumberToString);
+            const iface = new ethers.utils.Interface(abi);
+            const deployInfo = iface.deployFunction;
+            const txData = deployInfo.encode(bytecode, []);
+            const web3Wrapper = new web3_wrapper_1.Web3Wrapper(provider);
+            const txDataWithDefaults = yield base_contract_1.BaseContract._applyDefaultsToContractTxDataAsync(Object.assign({ data: txData }, txDefaults), web3Wrapper.estimateGasAsync.bind(web3Wrapper));
+            const txHash = yield web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+            utils_1.logUtils.log(`transactionHash: ${txHash}`);
+            const txReceipt = yield web3Wrapper.awaitTransactionSuccessAsync(txHash);
+            utils_1.logUtils.log(`WyvernAtomicizer successfully deployed at ${txReceipt.contractAddress}`);
+            const contractInstance = new WyvernAtomicizerContract(txReceipt.contractAddress, provider, txDefaults, logDecodeDependencies);
+            contractInstance.constructorArgs = [];
+            return contractInstance;
+        });
+    }
+    /**
+     * @returns      The contract ABI
+     */
+    static ABI() {
+        const abi = [
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'addrs',
+                        type: 'address[]',
+                    },
+                    {
+                        name: 'values',
+                        type: 'uint256[]',
+                    },
+                    {
+                        name: 'calldataLengths',
+                        type: 'uint256[]',
+                    },
+                    {
+                        name: 'calldatas',
+                        type: 'bytes',
+                    },
+                ],
+                name: 'atomicize',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+        ];
+        return abi;
+    }
+    static _deployLibrariesAsync(artifact, libraryArtifacts, web3Wrapper, txDefaults, libraryAddresses = {}) {
+        return __awaiter(this, void 0, void 0, function* () {
+            const links = artifact.compilerOutput.evm.bytecode.linkReferences;
+            // Go through all linked libraries, recursively deploying them if necessary.
+            for (const link of Object.values(links)) {
+                for (const libraryName of Object.keys(link)) {
+                    if (!libraryAddresses[libraryName]) {
+                        // Library not yet deployed.
+                        const libraryArtifact = libraryArtifacts[libraryName];
+                        if (!libraryArtifact) {
+                            throw new Error(`Missing artifact for linked library "${libraryName}"`);
                         }
-                    });
+                        // Deploy any dependent libraries used by this library.
+                        yield WyvernAtomicizerContract._deployLibrariesAsync(libraryArtifact, libraryArtifacts, web3Wrapper, txDefaults, libraryAddresses);
+                        // Deploy this library.
+                        const linkedLibraryBytecode = (0, base_contract_1.linkLibrariesInBytecode)(libraryArtifact, libraryAddresses);
+                        const txDataWithDefaults = yield base_contract_1.BaseContract._applyDefaultsToContractTxDataAsync(Object.assign({ data: linkedLibraryBytecode }, txDefaults), web3Wrapper.estimateGasAsync.bind(web3Wrapper));
+                        const txHash = yield web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                        utils_1.logUtils.log(`transactionHash: ${txHash}`);
+                        const { contractAddress } = yield web3Wrapper.awaitTransactionSuccessAsync(txHash);
+                        utils_1.logUtils.log(`${libraryArtifact.contractName} successfully deployed at ${contractAddress}`);
+                        libraryAddresses[libraryArtifact.contractName] = contractAddress;
+                    }
+                }
+            }
+            return libraryAddresses;
+        });
+    }
+    getFunctionSignature(methodName) {
+        const index = this._methodABIIndex[methodName];
+        const methodAbi = WyvernAtomicizerContract.ABI()[index]; // tslint:disable-line:no-unnecessary-type-assertion
+        const functionSignature = (0, base_contract_1.methodAbiToFunctionSignature)(methodAbi);
+        return functionSignature;
+    }
+    getABIDecodedTransactionData(methodName, callData) {
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = this;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        const abiDecodedCallData = abiEncoder.strictDecode(callData);
+        return abiDecodedCallData;
+    }
+    getABIDecodedReturnData(methodName, callData) {
+        if (this._encoderOverrides.decodeOutput) {
+            return this._encoderOverrides.decodeOutput(methodName, callData);
+        }
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = this;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        const abiDecodedCallData = abiEncoder.strictDecodeReturnValue(callData);
+        return abiDecodedCallData;
+    }
+    getSelector(methodName) {
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = this;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        return abiEncoder.getSelector();
+    }
+    atomicize(addrs, values, calldataLengths, calldatas) {
+        const self = this;
+        assert_1.assert.isArray('addrs', addrs);
+        assert_1.assert.isArray('values', values);
+        assert_1.assert.isArray('calldataLengths', calldataLengths);
+        assert_1.assert.isString('calldatas', calldatas);
+        const functionSignature = 'atomicize(address[],uint256[],uint256[],bytes)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
                 });
             },
-            estimateGasAsync: function (addrs_0, values_1, calldataLengths_2, calldatas_3, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.atomicize.estimateGas, self.web3ContractInstance)(addrs_0, values_1, calldataLengths_2, calldatas_3, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
                 });
             },
-            getABIEncodedTransactionData: function (addrs_0, values_1, calldataLengths_2, calldatas_3) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.atomicize.getData(addrs_0, values_1, calldataLengths_2, calldatas_3);
-                return abiEncodedTransactionData;
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [addrs,
+                    values,
+                    calldataLengths,
+                    calldatas
+                ]);
             },
         };
-        utils_1.classUtils.bindAll(_this, ['web3ContractInstance', 'defaults']);
-        return _this;
     }
-    return WyvernAtomicizerContract;
-}(base_contract_1.BaseContract)); // tslint:disable:max-file-line-count
+}
 exports.WyvernAtomicizerContract = WyvernAtomicizerContract;
+WyvernAtomicizerContract.contractName = 'WyvernAtomicizer';
+// tslint:disable:max-file-line-count
+// tslint:enable:no-unbound-method no-parameter-reassignment no-consecutive-blank-lines ordered-imports align
+// tslint:enable:trailing-comma whitespace no-trailing-whitespace
 //# sourceMappingURL=wyvern_atomicizer.js.map
