@@ -1,519 +1,961 @@
 "use strict";
-var __extends = (this && this.__extends) || (function () {
-    var extendStatics = Object.setPrototypeOf ||
-        ({ __proto__: [] } instanceof Array && function (d, b) { d.__proto__ = b; }) ||
-        function (d, b) { for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p]; };
-    return function (d, b) {
-        extendStatics(d, b);
-        function __() { this.constructor = d; }
-        d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
-    };
-})();
 var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
     return new (P || (P = Promise))(function (resolve, reject) {
         function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
         function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
-        function step(result) { result.done ? resolve(result.value) : new P(function (resolve) { resolve(result.value); }).then(fulfilled, rejected); }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-var __generator = (this && this.__generator) || function (thisArg, body) {
-    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
-    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
-    function verb(n) { return function (v) { return step([n, v]); }; }
-    function step(op) {
-        if (f) throw new TypeError("Generator is already executing.");
-        while (_) try {
-            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
-            if (y = 0, t) op = [op[0] & 2, t.value];
-            switch (op[0]) {
-                case 0: case 1: t = op; break;
-                case 4: _.label++; return { value: op[1], done: false };
-                case 5: _.label++; y = op[1]; op = [0]; continue;
-                case 7: op = _.ops.pop(); _.trys.pop(); continue;
-                default:
-                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
-                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
-                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
-                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
-                    if (t[2]) _.ops.pop();
-                    _.trys.pop(); continue;
-            }
-            op = body.call(thisArg, _);
-        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
-        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
-    }
-};
 Object.defineProperty(exports, "__esModule", { value: true });
-var utils_1 = require("@0xproject/utils");
-var base_contract_1 = require("./base_contract");
-var WyvernProxyRegistryContract = /** @class */ (function (_super) {
-    __extends(WyvernProxyRegistryContract, _super);
-    function WyvernProxyRegistryContract(web3ContractInstance, defaults) {
-        var _this = _super.call(this, web3ContractInstance, defaults) || this;
-        _this.name = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.name.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.initialAddressSet = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.initialAddressSet.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.endGrantAuthentication = {
-            sendTransactionAsync: function (addr_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.endGrantAuthentication.estimateGasAsync.bind(self, addr_0))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.endGrantAuthentication, self.web3ContractInstance)(addr_0, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (addr_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.endGrantAuthentication.estimateGas, self.web3ContractInstance)(addr_0, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (addr_0) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.endGrantAuthentication.getData(addr_0);
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.revokeAuthentication = {
-            sendTransactionAsync: function (addr_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.revokeAuthentication.estimateGasAsync.bind(self, addr_0))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.revokeAuthentication, self.web3ContractInstance)(addr_0, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (addr_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.revokeAuthentication.estimateGas, self.web3ContractInstance)(addr_0, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (addr_0) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.revokeAuthentication.getData(addr_0);
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.pending = {
-            callAsync: function (index_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.pending.call, self.web3ContractInstance)(index_0, txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.contracts = {
-            callAsync: function (index_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.contracts.call, self.web3ContractInstance)(index_0, txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.renounceOwnership = {
-            sendTransactionAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.renounceOwnership.estimateGasAsync.bind(self))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.renounceOwnership, self.web3ContractInstance)(txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.renounceOwnership.estimateGas, self.web3ContractInstance)(txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function () {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.renounceOwnership.getData();
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.owner = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.owner.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.delegateProxyImplementation = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.delegateProxyImplementation.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.proxies = {
-            callAsync: function (index_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.proxies.call, self.web3ContractInstance)(index_0, txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.startGrantAuthentication = {
-            sendTransactionAsync: function (addr_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.startGrantAuthentication.estimateGasAsync.bind(self, addr_0))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.startGrantAuthentication, self.web3ContractInstance)(addr_0, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (addr_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.startGrantAuthentication.estimateGas, self.web3ContractInstance)(addr_0, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (addr_0) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.startGrantAuthentication.getData(addr_0);
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.registerProxy = {
-            sendTransactionAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.registerProxy.estimateGasAsync.bind(self))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.registerProxy, self.web3ContractInstance)(txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.registerProxy.estimateGas, self.web3ContractInstance)(txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function () {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.registerProxy.getData();
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.DELAY_PERIOD = {
-            callAsync: function (txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, result;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.DELAY_PERIOD.call, self.web3ContractInstance)(txData)];
-                            case 1:
-                                result = _a.sent();
-                                return [2 /*return*/, result];
-                        }
-                    });
-                });
-            },
-        };
-        _this.transferOwnership = {
-            sendTransactionAsync: function (newOwner_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.transferOwnership.estimateGasAsync.bind(self, newOwner_0))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.transferOwnership, self.web3ContractInstance)(newOwner_0, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (newOwner_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.transferOwnership.estimateGas, self.web3ContractInstance)(newOwner_0, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (newOwner_0) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.transferOwnership.getData(newOwner_0);
-                return abiEncodedTransactionData;
-            },
-        };
-        _this.grantInitialAuthentication = {
-            sendTransactionAsync: function (authAddress_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, txHash;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData, self.grantInitialAuthentication.estimateGasAsync.bind(self, authAddress_0))];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.grantInitialAuthentication, self.web3ContractInstance)(authAddress_0, txDataWithDefaults)];
-                            case 2:
-                                txHash = _a.sent();
-                                return [2 /*return*/, txHash];
-                        }
-                    });
-                });
-            },
-            estimateGasAsync: function (authAddress_0, txData) {
-                if (txData === void 0) { txData = {}; }
-                return __awaiter(this, void 0, void 0, function () {
-                    var self, txDataWithDefaults, gas;
-                    return __generator(this, function (_a) {
-                        switch (_a.label) {
-                            case 0:
-                                self = this;
-                                return [4 /*yield*/, self.applyDefaultsToTxDataAsync(txData)];
-                            case 1:
-                                txDataWithDefaults = _a.sent();
-                                return [4 /*yield*/, utils_1.promisify(self.web3ContractInstance.grantInitialAuthentication.estimateGas, self.web3ContractInstance)(authAddress_0, txDataWithDefaults)];
-                            case 2:
-                                gas = _a.sent();
-                                return [2 /*return*/, gas];
-                        }
-                    });
-                });
-            },
-            getABIEncodedTransactionData: function (authAddress_0) {
-                var self = this;
-                var abiEncodedTransactionData = self.web3ContractInstance.grantInitialAuthentication.getData(authAddress_0);
-                return abiEncodedTransactionData;
-            },
-        };
-        utils_1.classUtils.bindAll(_this, ['web3ContractInstance', 'defaults']);
-        return _this;
+exports.WyvernProxyRegistryContract = exports.WyvernProxyRegistryEvents = void 0;
+// tslint:disable:no-consecutive-blank-lines ordered-imports align trailing-comma enum-naming
+// tslint:disable:whitespace no-unbound-method no-trailing-whitespace
+// tslint:disable:no-unused-variable
+const base_contract_1 = require("@0x/base-contract");
+const json_schemas_1 = require("@0x/json-schemas");
+const utils_1 = require("@0x/utils");
+const web3_wrapper_1 = require("@0x/web3-wrapper");
+const assert_1 = require("@0x/assert");
+const ethers = require("ethers");
+var WyvernProxyRegistryEvents;
+(function (WyvernProxyRegistryEvents) {
+    WyvernProxyRegistryEvents["OwnershipRenounced"] = "OwnershipRenounced";
+    WyvernProxyRegistryEvents["OwnershipTransferred"] = "OwnershipTransferred";
+})(WyvernProxyRegistryEvents = exports.WyvernProxyRegistryEvents || (exports.WyvernProxyRegistryEvents = {}));
+/* istanbul ignore next */
+// tslint:disable:array-type
+// tslint:disable:no-parameter-reassignment
+// tslint:disable-next-line:class-name
+class WyvernProxyRegistryContract extends base_contract_1.BaseContract {
+    constructor(address, supportedProvider, txDefaults, logDecodeDependencies, deployedBytecode = WyvernProxyRegistryContract.deployedBytecode, encoderOverrides) {
+        super('WyvernProxyRegistry', WyvernProxyRegistryContract.ABI(), address, supportedProvider, txDefaults, logDecodeDependencies, deployedBytecode, encoderOverrides);
+        this._methodABIIndex = {};
+        utils_1.classUtils.bindAll(this, ['_abiEncoderByFunctionSignature', 'address', '_web3Wrapper']);
+        this._subscriptionManager = new base_contract_1.SubscriptionManager(WyvernProxyRegistryContract.ABI(), this._web3Wrapper);
+        WyvernProxyRegistryContract.ABI().forEach((item, index) => {
+            if (item.type === 'function') {
+                const methodAbi = item;
+                this._methodABIIndex[methodAbi.name] = index;
+            }
+        });
     }
-    return WyvernProxyRegistryContract;
-}(base_contract_1.BaseContract)); // tslint:disable:max-file-line-count
+    static deployFrom0xArtifactAsync(artifact, supportedProvider, txDefaults, logDecodeDependencies) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.doesConformToSchema('txDefaults', txDefaults, json_schemas_1.schemas.txDataSchema);
+            if (artifact.compilerOutput === undefined) {
+                throw new Error('Compiler output not found in the artifact file');
+            }
+            const provider = utils_1.providerUtils.standardizeOrThrow(supportedProvider);
+            const bytecode = artifact.compilerOutput.evm.bytecode.object;
+            const abi = artifact.compilerOutput.abi;
+            const logDecodeDependenciesAbiOnly = {};
+            if (Object.keys(logDecodeDependencies) !== undefined) {
+                for (const key of Object.keys(logDecodeDependencies)) {
+                    logDecodeDependenciesAbiOnly[key] = logDecodeDependencies[key].compilerOutput.abi;
+                }
+            }
+            return WyvernProxyRegistryContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly);
+        });
+    }
+    static deployWithLibrariesFrom0xArtifactAsync(artifact, libraryArtifacts, supportedProvider, txDefaults, logDecodeDependencies) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.doesConformToSchema('txDefaults', txDefaults, json_schemas_1.schemas.txDataSchema);
+            if (artifact.compilerOutput === undefined) {
+                throw new Error('Compiler output not found in the artifact file');
+            }
+            const provider = utils_1.providerUtils.standardizeOrThrow(supportedProvider);
+            const abi = artifact.compilerOutput.abi;
+            const logDecodeDependenciesAbiOnly = {};
+            if (Object.keys(logDecodeDependencies) !== undefined) {
+                for (const key of Object.keys(logDecodeDependencies)) {
+                    logDecodeDependenciesAbiOnly[key] = logDecodeDependencies[key].compilerOutput.abi;
+                }
+            }
+            const libraryAddresses = yield WyvernProxyRegistryContract._deployLibrariesAsync(artifact, libraryArtifacts, new web3_wrapper_1.Web3Wrapper(provider), txDefaults);
+            const bytecode = (0, base_contract_1.linkLibrariesInBytecode)(artifact, libraryAddresses);
+            return WyvernProxyRegistryContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly);
+        });
+    }
+    static deployAsync(bytecode, abi, supportedProvider, txDefaults, logDecodeDependencies) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.isHexString('bytecode', bytecode);
+            assert_1.assert.doesConformToSchema('txDefaults', txDefaults, json_schemas_1.schemas.txDataSchema);
+            const provider = utils_1.providerUtils.standardizeOrThrow(supportedProvider);
+            const constructorAbi = base_contract_1.BaseContract._lookupConstructorAbi(abi);
+            [] = base_contract_1.BaseContract._formatABIDataItemList(constructorAbi.inputs, [], base_contract_1.BaseContract._bigNumberToString);
+            const iface = new ethers.utils.Interface(abi);
+            const deployInfo = iface.deployFunction;
+            const txData = deployInfo.encode(bytecode, []);
+            const web3Wrapper = new web3_wrapper_1.Web3Wrapper(provider);
+            const txDataWithDefaults = yield base_contract_1.BaseContract._applyDefaultsToContractTxDataAsync(Object.assign({ data: txData }, txDefaults), web3Wrapper.estimateGasAsync.bind(web3Wrapper));
+            const txHash = yield web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+            utils_1.logUtils.log(`transactionHash: ${txHash}`);
+            const txReceipt = yield web3Wrapper.awaitTransactionSuccessAsync(txHash);
+            utils_1.logUtils.log(`WyvernProxyRegistry successfully deployed at ${txReceipt.contractAddress}`);
+            const contractInstance = new WyvernProxyRegistryContract(txReceipt.contractAddress, provider, txDefaults, logDecodeDependencies);
+            contractInstance.constructorArgs = [];
+            return contractInstance;
+        });
+    }
+    /**
+     * @returns      The contract ABI
+     */
+    static ABI() {
+        const abi = [
+            {
+                constant: true,
+                inputs: [],
+                name: 'name',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'string',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [],
+                name: 'initialAddressSet',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'bool',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'addr',
+                        type: 'address',
+                    },
+                ],
+                name: 'endGrantAuthentication',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'addr',
+                        type: 'address',
+                    },
+                ],
+                name: 'revokeAuthentication',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [
+                    {
+                        name: 'index_0',
+                        type: 'address',
+                    },
+                ],
+                name: 'pending',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'uint256',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [
+                    {
+                        name: 'index_0',
+                        type: 'address',
+                    },
+                ],
+                name: 'contracts',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'bool',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [],
+                name: 'renounceOwnership',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [],
+                name: 'owner',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [],
+                name: 'delegateProxyImplementation',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [
+                    {
+                        name: 'index_0',
+                        type: 'address',
+                    },
+                ],
+                name: 'proxies',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'addr',
+                        type: 'address',
+                    },
+                ],
+                name: 'startGrantAuthentication',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [],
+                name: 'registerProxy',
+                outputs: [
+                    {
+                        name: 'proxy',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                constant: true,
+                inputs: [],
+                name: 'DELAY_PERIOD',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'uint256',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'newOwner',
+                        type: 'address',
+                    },
+                ],
+                name: 'transferOwnership',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            {
+                inputs: [],
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'constructor',
+            },
+            {
+                anonymous: false,
+                inputs: [
+                    {
+                        name: 'previousOwner',
+                        type: 'address',
+                        indexed: true,
+                    },
+                ],
+                name: 'OwnershipRenounced',
+                outputs: [],
+                type: 'event',
+            },
+            {
+                anonymous: false,
+                inputs: [
+                    {
+                        name: 'previousOwner',
+                        type: 'address',
+                        indexed: true,
+                    },
+                    {
+                        name: 'newOwner',
+                        type: 'address',
+                        indexed: true,
+                    },
+                ],
+                name: 'OwnershipTransferred',
+                outputs: [],
+                type: 'event',
+            },
+            {
+                constant: false,
+                inputs: [
+                    {
+                        name: 'authAddress',
+                        type: 'address',
+                    },
+                ],
+                name: 'grantInitialAuthentication',
+                outputs: [],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+        ];
+        return abi;
+    }
+    static _deployLibrariesAsync(artifact, libraryArtifacts, web3Wrapper, txDefaults, libraryAddresses = {}) {
+        return __awaiter(this, void 0, void 0, function* () {
+            const links = artifact.compilerOutput.evm.bytecode.linkReferences;
+            // Go through all linked libraries, recursively deploying them if necessary.
+            for (const link of Object.values(links)) {
+                for (const libraryName of Object.keys(link)) {
+                    if (!libraryAddresses[libraryName]) {
+                        // Library not yet deployed.
+                        const libraryArtifact = libraryArtifacts[libraryName];
+                        if (!libraryArtifact) {
+                            throw new Error(`Missing artifact for linked library "${libraryName}"`);
+                        }
+                        // Deploy any dependent libraries used by this library.
+                        yield WyvernProxyRegistryContract._deployLibrariesAsync(libraryArtifact, libraryArtifacts, web3Wrapper, txDefaults, libraryAddresses);
+                        // Deploy this library.
+                        const linkedLibraryBytecode = (0, base_contract_1.linkLibrariesInBytecode)(libraryArtifact, libraryAddresses);
+                        const txDataWithDefaults = yield base_contract_1.BaseContract._applyDefaultsToContractTxDataAsync(Object.assign({ data: linkedLibraryBytecode }, txDefaults), web3Wrapper.estimateGasAsync.bind(web3Wrapper));
+                        const txHash = yield web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                        utils_1.logUtils.log(`transactionHash: ${txHash}`);
+                        const { contractAddress } = yield web3Wrapper.awaitTransactionSuccessAsync(txHash);
+                        utils_1.logUtils.log(`${libraryArtifact.contractName} successfully deployed at ${contractAddress}`);
+                        libraryAddresses[libraryArtifact.contractName] = contractAddress;
+                    }
+                }
+            }
+            return libraryAddresses;
+        });
+    }
+    getFunctionSignature(methodName) {
+        const index = this._methodABIIndex[methodName];
+        const methodAbi = WyvernProxyRegistryContract.ABI()[index]; // tslint:disable-line:no-unnecessary-type-assertion
+        const functionSignature = (0, base_contract_1.methodAbiToFunctionSignature)(methodAbi);
+        return functionSignature;
+    }
+    getABIDecodedTransactionData(methodName, callData) {
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = this;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        const abiDecodedCallData = abiEncoder.strictDecode(callData);
+        return abiDecodedCallData;
+    }
+    getABIDecodedReturnData(methodName, callData) {
+        if (this._encoderOverrides.decodeOutput) {
+            return this._encoderOverrides.decodeOutput(methodName, callData);
+        }
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = this;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        const abiDecodedCallData = abiEncoder.strictDecodeReturnValue(callData);
+        return abiDecodedCallData;
+    }
+    getSelector(methodName) {
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = this;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        return abiEncoder.getSelector();
+    }
+    name() {
+        const self = this;
+        const functionSignature = 'name()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    initialAddressSet() {
+        const self = this;
+        const functionSignature = 'initialAddressSet()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    endGrantAuthentication(addr) {
+        const self = this;
+        assert_1.assert.isString('addr', addr);
+        const functionSignature = 'endGrantAuthentication(address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [addr.toLowerCase()
+                ]);
+            },
+        };
+    }
+    revokeAuthentication(addr) {
+        const self = this;
+        assert_1.assert.isString('addr', addr);
+        const functionSignature = 'revokeAuthentication(address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [addr.toLowerCase()
+                ]);
+            },
+        };
+    }
+    pending(index_0) {
+        const self = this;
+        assert_1.assert.isString('index_0', index_0);
+        const functionSignature = 'pending(address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [index_0.toLowerCase()
+                ]);
+            },
+        };
+    }
+    contracts(index_0) {
+        const self = this;
+        assert_1.assert.isString('index_0', index_0);
+        const functionSignature = 'contracts(address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [index_0.toLowerCase()
+                ]);
+            },
+        };
+    }
+    renounceOwnership() {
+        const self = this;
+        const functionSignature = 'renounceOwnership()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    owner() {
+        const self = this;
+        const functionSignature = 'owner()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    delegateProxyImplementation() {
+        const self = this;
+        const functionSignature = 'delegateProxyImplementation()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    proxies(index_0) {
+        const self = this;
+        assert_1.assert.isString('index_0', index_0);
+        const functionSignature = 'proxies(address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [index_0.toLowerCase()
+                ]);
+            },
+        };
+    }
+    startGrantAuthentication(addr) {
+        const self = this;
+        assert_1.assert.isString('addr', addr);
+        const functionSignature = 'startGrantAuthentication(address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [addr.toLowerCase()
+                ]);
+            },
+        };
+    }
+    registerProxy() {
+        const self = this;
+        const functionSignature = 'registerProxy()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    DELAY_PERIOD() {
+        const self = this;
+        const functionSignature = 'DELAY_PERIOD()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    transferOwnership(newOwner) {
+        const self = this;
+        assert_1.assert.isString('newOwner', newOwner);
+        const functionSignature = 'transferOwnership(address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [newOwner.toLowerCase()
+                ]);
+            },
+        };
+    }
+    grantInitialAuthentication(authAddress) {
+        const self = this;
+        assert_1.assert.isString('authAddress', authAddress);
+        const functionSignature = 'grantInitialAuthentication(address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+        return {
+            selector,
+            sendTransactionAsync(txData, opts = { shouldValidate: true }) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData), this.estimateGasAsync.bind(this));
+                    if (opts.shouldValidate !== false) {
+                        yield this.callAsync(txDataWithDefaults);
+                    }
+                    return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                });
+            },
+            awaitTransactionSuccessAsync(txData, opts = { shouldValidate: true }) {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            estimateGasAsync(txData) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+                });
+            },
+            createAccessListAsync(txData, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    const txDataWithDefaults = yield self._applyDefaultsToTxDataAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, txData));
+                    return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+                });
+            },
+            callAsync(callData = {}, defaultBlock) {
+                return __awaiter(this, void 0, void 0, function* () {
+                    base_contract_1.BaseContract._assertCallParams(callData, defaultBlock);
+                    const rawCallResult = yield self._performCallAsync(Object.assign({ data: this.getABIEncodedTransactionData() }, callData), defaultBlock);
+                    const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                    base_contract_1.BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                    return abiEncoder.strictDecodeReturnValue(rawCallResult);
+                });
+            },
+            getABIEncodedTransactionData() {
+                return self._strictEncodeArguments(functionSignature, [authAddress.toLowerCase()
+                ]);
+            },
+        };
+    }
+    /**
+     * Subscribe to an event type emitted by the WyvernProxyRegistry contract.
+     * @param eventName The WyvernProxyRegistry contract event you would like to subscribe to.
+     * @param indexFilterValues An object where the keys are indexed args returned by the event and
+     * the value is the value you are interested in. E.g `{maker: aUserAddressHex}`
+     * @param callback Callback that gets called when a log is added/removed
+     * @param isVerbose Enable verbose subscription warnings (e.g recoverable network issues encountered)
+     * @return Subscription token used later to unsubscribe
+     */
+    subscribe(eventName, indexFilterValues, callback, isVerbose = false, blockPollingIntervalMs) {
+        assert_1.assert.doesBelongToStringEnum('eventName', eventName, WyvernProxyRegistryEvents);
+        assert_1.assert.doesConformToSchema('indexFilterValues', indexFilterValues, json_schemas_1.schemas.indexFilterValuesSchema);
+        assert_1.assert.isFunction('callback', callback);
+        const subscriptionToken = this._subscriptionManager.subscribe(this.address, eventName, indexFilterValues, WyvernProxyRegistryContract.ABI(), callback, isVerbose, blockPollingIntervalMs);
+        return subscriptionToken;
+    }
+    /**
+     * Cancel a subscription
+     * @param subscriptionToken Subscription token returned by `subscribe()`
+     */
+    unsubscribe(subscriptionToken) {
+        this._subscriptionManager.unsubscribe(subscriptionToken);
+    }
+    /**
+     * Cancels all existing subscriptions
+     */
+    unsubscribeAll() {
+        this._subscriptionManager.unsubscribeAll();
+    }
+    /**
+     * Gets historical logs without creating a subscription
+     * @param eventName The WyvernProxyRegistry contract event you would like to subscribe to.
+     * @param blockRange Block range to get logs from.
+     * @param indexFilterValues An object where the keys are indexed args returned by the event and
+     * the value is the value you are interested in. E.g `{_from: aUserAddressHex}`
+     * @return Array of logs that match the parameters
+     */
+    getLogsAsync(eventName, blockRange, indexFilterValues) {
+        return __awaiter(this, void 0, void 0, function* () {
+            assert_1.assert.doesBelongToStringEnum('eventName', eventName, WyvernProxyRegistryEvents);
+            assert_1.assert.doesConformToSchema('blockRange', blockRange, json_schemas_1.schemas.blockRangeSchema);
+            assert_1.assert.doesConformToSchema('indexFilterValues', indexFilterValues, json_schemas_1.schemas.indexFilterValuesSchema);
+            const logs = yield this._subscriptionManager.getLogsAsync(this.address, eventName, blockRange, indexFilterValues, WyvernProxyRegistryContract.ABI());
+            return logs;
+        });
+    }
+}
 exports.WyvernProxyRegistryContract = WyvernProxyRegistryContract;
+WyvernProxyRegistryContract.contractName = 'WyvernProxyRegistry';
+// tslint:disable:max-file-line-count
+// tslint:enable:no-unbound-method no-parameter-reassignment no-consecutive-blank-lines ordered-imports align
+// tslint:enable:trailing-comma whitespace no-trailing-whitespace
 //# sourceMappingURL=wyvern_proxy_registry.js.map
