@@ -1,5 +1,6 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
+exports.WyvernProtocol = void 0;
 var wyvernProtocol_1 = require("./wyvernProtocol");
-exports.WyvernProtocol = wyvernProtocol_1.WyvernProtocol;
+Object.defineProperty(exports, "WyvernProtocol", { enumerable: true, get: function () { return wyvernProtocol_1.WyvernProtocol; } });
 //# sourceMappingURL=index.js.map
