@@ -1,8 +1,9 @@
 "use strict";
 Object.defineProperty(exports, "__esModule", { value: true });
-var json_schemas_1 = require("@0xproject/json-schemas");
-var data = { type: 'string', pattern: '^0x[0-9a-f]*$' };
-var orderSchema = {
+exports.schemas = void 0;
+const json_schemas_1 = require("@0x/json-schemas");
+const data = { type: 'string', pattern: '^0x[0-9a-f]*$' };
+const orderSchema = {
     id: '/Order',
     properties: {
         exchange: { $ref: '/Address' },
@@ -56,7 +57,7 @@ var orderSchema = {
     ],
     type: 'object',
 };
-var signedOrderSchema = {
+const signedOrderSchema = {
     id: '/SignedOrder',
     allOf: [
         { $ref: '/Order' },
@@ -74,7 +75,7 @@ exports.schemas = {
     ecSignatureSchema: json_schemas_1.schemas.ecSignatureSchema,
     ecSignatureParameterSchema: json_schemas_1.schemas.ecSignatureParameterSchema,
     orderHashSchema: json_schemas_1.schemas.orderHashSchema,
-    orderSchema: orderSchema,
-    signedOrderSchema: signedOrderSchema,
+    orderSchema,
+    signedOrderSchema,
 };
 //# sourceMappingURL=schemas.js.map
