@@ -1,10 +1,10 @@
-import * as Web3 from 'web3';
+import { AbiDefinition, LogEntry } from 'ethereum-types';
 import { ContractEventArgs, LogWithDecodedArgs, RawLog } from '../types';
 export declare class AbiDecoder {
     private _savedABIs;
     private _methodIds;
     private static _padZeros;
-    constructor(abiArrays: Web3.AbiDefinition[][]);
-    tryToDecodeLogOrNoop<ArgsType extends ContractEventArgs>(log: Web3.LogEntry): LogWithDecodedArgs<ArgsType> | RawLog;
+    constructor(abiArrays: AbiDefinition[][]);
+    tryToDecodeLogOrNoop<ArgsType extends ContractEventArgs>(log: LogEntry): LogWithDecodedArgs<ArgsType> | RawLog;
     private _addABI;
 }
