@@ -1,12 +1,12 @@
-import { Schema } from '@0xproject/json-schemas';
-import { BigNumber } from '@0xproject/utils';
-import { Web3Wrapper } from '@0xproject/web3-wrapper';
+import { BigNumber } from '@0x/utils';
+import { Web3Wrapper } from '@0x/web3-wrapper';
 import { ECSignature } from '../types';
 export declare const assert: {
     isValidSignature(orderHash: string, ecSignature: ECSignature, signerAddress: string): void;
     isSenderAddressAsync(variableName: string, senderAddressHex: string, web3Wrapper: Web3Wrapper): Promise<void>;
     isUserAddressAvailableAsync(web3Wrapper: Web3Wrapper): Promise<void>;
     isBigNumber(variableName: string, value: BigNumber): void;
+    isNumberLike(variableName: string, value: number | BigNumber): void;
     isValidBaseUnitAmount(variableName: string, value: BigNumber): void;
     isString(variableName: string, value: string): void;
     isFunction(variableName: string, value: any): void;
@@ -15,11 +15,15 @@ export declare const assert: {
     doesBelongToStringEnum(variableName: string, value: string, stringEnum: any): void;
     hasAtMostOneUniqueValue(value: any[], errMsg: string): void;
     isNumber(variableName: string, value: number): void;
+    isNumberOrBigNumber(variableName: string, value: any): void;
     isBoolean(variableName: string, value: boolean): void;
     isWeb3Provider(variableName: string, value: any): void;
-    doesConformToSchema(variableName: string, value: any, schema: Schema): void;
-    isHttpUrl(variableName: string, value: any): void;
+    doesConformToSchema(variableName: string, value: any, schema: object, subSchemas?: object[] | undefined): void;
+    doesMatchRegex(variableName: string, value: string, regex: RegExp): void;
+    isWebUri(variableName: string, value: any): void;
     isUri(variableName: string, value: any): void;
+    isBlockParam(variableName: string, value: any): void;
+    isArray(variableName: string, value: any): void;
     assert(condition: boolean, message: string): void;
     typeAssertionMessage(variableName: string, type: string, value: any): string;
 };
