@@ -1,12 +1,12 @@
 export declare const atomicizerABI: {
-    'constant': boolean;
-    'inputs': {
-        'name': string;
-        'type': string;
+    constant: boolean;
+    inputs: {
+        name: string;
+        type: string;
     }[];
-    'name': string;
-    'outputs': never[];
-    'payable': boolean;
-    'stateMutability': string;
-    'type': string;
+    name: string;
+    outputs: never[];
+    payable: boolean;
+    stateMutability: string;
+    type: string;
 }[];
