@@ -1,51 +1,51 @@
 export declare const deployed: {
-    'rinkeby': {
-        'TestToken': string;
-        'TestDAO': string;
-        'Migrations': string;
-        'MerkleProof': string;
-        'WyvernToken': string;
-        'WyvernDAO': string;
-        'WyvernRegistry': string;
-        'SaleKindInterface': string;
-        'WyvernExchange': string;
-        'DirectEscrowProvider': string;
-        'WyvernProxyRegistry': string;
-        'TestStatic': string;
-        'WyvernDAOProxy': string;
-        'WyvernAtomicizer': string;
-        'WyvernTokenTransferProxy': string;
+    rinkeby: {
+        TestToken: string;
+        TestDAO: string;
+        Migrations: string;
+        MerkleProof: string;
+        WyvernToken: string;
+        WyvernDAO: string;
+        WyvernRegistry: string;
+        SaleKindInterface: string;
+        WyvernExchange: string;
+        DirectEscrowProvider: string;
+        WyvernProxyRegistry: string;
+        TestStatic: string;
+        WyvernDAOProxy: string;
+        WyvernAtomicizer: string;
+        WyvernTokenTransferProxy: string;
     };
-    'development': {
-        'Migrations': string;
-        'TestToken': string;
-        'TestDAO': string;
-        'MerkleProof': string;
-        'WyvernToken': string;
-        'WyvernDAO': string;
-        'WyvernRegistry': string;
-        'SaleKindInterface': string;
-        'WyvernExchange': string;
-        'DirectEscrowProvider': string;
-        'WyvernProxyRegistry': string;
-        'TestStatic': string;
-        'WyvernDAOProxy': string;
-        'WyvernAtomicizer': string;
-        'WyvernTokenTransferProxy': string;
+    development: {
+        Migrations: string;
+        TestToken: string;
+        TestDAO: string;
+        MerkleProof: string;
+        WyvernToken: string;
+        WyvernDAO: string;
+        WyvernRegistry: string;
+        SaleKindInterface: string;
+        WyvernExchange: string;
+        DirectEscrowProvider: string;
+        WyvernProxyRegistry: string;
+        TestStatic: string;
+        WyvernDAOProxy: string;
+        WyvernAtomicizer: string;
+        WyvernTokenTransferProxy: string;
     };
-    'main': {
-        'Migrations': string;
-        'MerkleProof': string;
-        'WyvernToken': string;
-        'WyvernDAO': string;
-        'TestToken': string;
-        'TestDAO': string;
-        'TestStatic': string;
-        'WyvernProxyRegistry': string;
-        'SaleKindInterface': string;
-        'WyvernExchange': string;
-        'WyvernDAOProxy': string;
-        'WyvernAtomicizer': string;
-        'WyvernTokenTransferProxy': string;
+    main: {
+        Migrations: string;
+        MerkleProof: string;
+        WyvernToken: string;
+        WyvernDAO: string;
+        TestToken: string;
+        TestDAO: string;
+        TestStatic: string;
+        WyvernProxyRegistry: string;
+        SaleKindInterface: string;
+        WyvernExchange: string;
+        WyvernDAOProxy: string;
+        WyvernAtomicizer: string;
+        WyvernTokenTransferProxy: string;
     };
 };
