@@ -1,40 +1,41 @@
 "use strict";
 /* Sourced from 0x.js */
 Object.defineProperty(exports, "__esModule", { value: true });
-var ethUtil = require("ethereumjs-util");
+exports.signatureUtils = void 0;
+const ethUtil = require("ethereumjs-util");
 exports.signatureUtils = {
-    isValidSignature: function (data, signature, signerAddress) {
-        var dataBuff = ethUtil.toBuffer(data);
+    isValidSignature(data, signature, signerAddress) {
+        const dataBuff = ethUtil.toBuffer(data);
         // const msgHashBuff = ethUtil.hashPersonalMessage(dataBuff);
-        var msgHashBuff = dataBuff;
+        const msgHashBuff = dataBuff;
         try {
-            var pubKey = ethUtil.ecrecover(msgHashBuff, signature.v, ethUtil.toBuffer(signature.r), ethUtil.toBuffer(signature.s));
-            var retrievedAddress = ethUtil.bufferToHex(ethUtil.pubToAddress(pubKey));
+            const pubKey = ethUtil.ecrecover(msgHashBuff, signature.v, ethUtil.toBuffer(signature.r), ethUtil.toBuffer(signature.s));
+            const retrievedAddress = ethUtil.bufferToHex(ethUtil.pubToAddress(pubKey));
             return retrievedAddress === signerAddress;
         }
         catch (err) {
             return false;
         }
     },
-    parseSignatureHexAsVRS: function (signatureHex) {
-        var signatureBuffer = ethUtil.toBuffer(signatureHex);
-        var v = signatureBuffer[0];
+    parseSignatureHexAsVRS(signatureHex) {
+        const signatureBuffer = ethUtil.toBuffer(signatureHex);
+        let v = signatureBuffer[0];
         if (v < 27) {
             v += 27;
         }
-        var r = signatureBuffer.slice(1, 33);
-        var s = signatureBuffer.slice(33, 65);
-        var ecSignature = {
-            v: v,
+        const r = signatureBuffer.slice(1, 33);
+        const s = signatureBuffer.slice(33, 65);
+        const ecSignature = {
+            v,
             r: ethUtil.bufferToHex(r),
             s: ethUtil.bufferToHex(s),
         };
         return ecSignature;
     },
-    parseSignatureHexAsRSV: function (signatureHex) {
-        var _a = ethUtil.fromRpcSig(signatureHex), v = _a.v, r = _a.r, s = _a.s;
-        var ecSignature = {
-            v: v,
+    parseSignatureHexAsRSV(signatureHex) {
+        const { v, r, s } = ethUtil.fromRpcSig(signatureHex);
+        const ecSignature = {
+            v,
             r: ethUtil.bufferToHex(r),
             s: ethUtil.bufferToHex(s),
         };
