@@ -1,404 +1,1140 @@
-/**
- * This file is auto-generated using abi-gen. Don't edit directly.
- * Templates can be found at https://github.com/0xProject/0x.js/tree/development/packages/abi-gen-templates.
- */
-// tslint:disable-next-line:no-unused-variable
-import { TxData, TxDataPayable } from '@0xproject/types';
-import { BigNumber, classUtils, promisify } from '@0xproject/utils';
-import * as Web3 from 'web3';
+// tslint:disable:no-consecutive-blank-lines ordered-imports align trailing-comma enum-naming
+// tslint:disable:whitespace no-unbound-method no-trailing-whitespace
+// tslint:disable:no-unused-variable
+import {
+    AwaitTransactionSuccessOpts,
+    EncoderOverrides,
+    ContractFunctionObj,
+    ContractTxFunctionObj,
+    SendTransactionOpts,
+    BaseContract,
+    SubscriptionManager,PromiseWithTransactionHash,
+    methodAbiToFunctionSignature,
+    linkLibrariesInBytecode,
+} from '@0x/base-contract';
+import { schemas } from '@0x/json-schemas';
+import {
+    BlockParam,
+    BlockParamLiteral,
+    BlockRange,
+    CallData,
+    ContractAbi,
+    ContractArtifact,
+    DecodedLogArgs,
+    LogWithDecodedArgs,
+    MethodAbi,
+    TransactionReceiptWithDecodedLogs,
+    TxData,
+    TxDataPayable,
+    TxAccessListWithGas,
+    SupportedProvider,
+} from 'ethereum-types';
+import { AbiEncoder, BigNumber, classUtils, EncodingRules, hexUtils, logUtils, providerUtils } from '@0x/utils';
+import { EventCallback, IndexedFilterValues, SimpleContractArtifact } from '@0x/types';
+import { Web3Wrapper } from '@0x/web3-wrapper';
+import { assert } from '@0x/assert';
+import * as ethers from 'ethers';
+// tslint:enable:no-unused-variable
 
-import {BaseContract} from './base_contract';
 
+export type AuthenticatedProxyEventArgs =
+    | AuthenticatedProxyRevokedEventArgs
+    | AuthenticatedProxyReceivedEtherEventArgs
+    | AuthenticatedProxyReceivedTokensEventArgs;
+
+export enum AuthenticatedProxyEvents {
+    Revoked = 'Revoked',
+    ReceivedEther = 'ReceivedEther',
+    ReceivedTokens = 'ReceivedTokens',
+}
+
+export interface AuthenticatedProxyRevokedEventArgs extends DecodedLogArgs {
+    revoked: boolean;
+}
+
+export interface AuthenticatedProxyReceivedEtherEventArgs extends DecodedLogArgs {
+    sender: string;
+    amount: BigNumber;
+}
+
+export interface AuthenticatedProxyReceivedTokensEventArgs extends DecodedLogArgs {
+    from: string;
+    value: BigNumber;
+    token: string;
+    extraData: string;
+}
+
+
+/* istanbul ignore next */
+// tslint:disable:array-type
+// tslint:disable:no-parameter-reassignment
+// tslint:disable-next-line:class-name
 export class AuthenticatedProxyContract extends BaseContract {
-    public proxyType = {
-        async callAsync(
-            txData: TxData = {},
-        ): Promise<BigNumber
-    > {
-            const self = this as AuthenticatedProxyContract;
-            const result = await promisify<BigNumber
-    >(
-                self.web3ContractInstance.proxyType.call,
-                self.web3ContractInstance,
-            )(
-                txData,
-            );
-            return result;
-        },
-    };
-    public user = {
-        async callAsync(
-            txData: TxData = {},
-        ): Promise<string
-    > {
-            const self = this as AuthenticatedProxyContract;
-            const result = await promisify<string
-    >(
-                self.web3ContractInstance.user.call,
-                self.web3ContractInstance,
-            )(
-                txData,
-            );
-            return result;
-        },
-    };
-    public implementation = {
-        async callAsync(
-            txData: TxData = {},
-        ): Promise<string
-    > {
-            const self = this as AuthenticatedProxyContract;
-            const result = await promisify<string
-    >(
-                self.web3ContractInstance.implementation.call,
-                self.web3ContractInstance,
-            )(
-                txData,
-            );
-            return result;
-        },
-    };
-    public revoked = {
-        async callAsync(
-            txData: TxData = {},
-        ): Promise<boolean
-    > {
-            const self = this as AuthenticatedProxyContract;
-            const result = await promisify<boolean
-    >(
-                self.web3ContractInstance.revoked.call,
-                self.web3ContractInstance,
-            )(
-                txData,
-            );
-            return result;
-        },
-    };
-    public upgradeabilityOwner = {
-        async callAsync(
-            txData: TxData = {},
-        ): Promise<string
-    > {
-            const self = this as AuthenticatedProxyContract;
-            const result = await promisify<string
-    >(
-                self.web3ContractInstance.upgradeabilityOwner.call,
-                self.web3ContractInstance,
-            )(
-                txData,
-            );
-            return result;
-        },
-    };
-    public registry = {
-        async callAsync(
-            txData: TxData = {},
-        ): Promise<string
-    > {
-            const self = this as AuthenticatedProxyContract;
-            const result = await promisify<string
-    >(
-                self.web3ContractInstance.registry.call,
-                self.web3ContractInstance,
-            )(
-                txData,
-            );
-            return result;
-        },
-    };
-    public receiveApproval = {
-        async sendTransactionAsync(
-            from_0: string,
-            value_1: BigNumber,
-            token_2: string,
-            extraData_3: string,
-            txData: TxData = {},
-        ): Promise<string> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-                self.receiveApproval.estimateGasAsync.bind(
-                    self,
-                    from_0,
-                    value_1,
-                    token_2,
-                    extraData_3,
-                ),
-            );
-            const txHash = await promisify<string>(
-                self.web3ContractInstance.receiveApproval, self.web3ContractInstance,
-            )(
-                from_0,
-                value_1,
-                token_2,
-                extraData_3,
-                txDataWithDefaults,
-            );
-            return txHash;
-        },
-        async estimateGasAsync(
-            from_0: string,
-            value_1: BigNumber,
-            token_2: string,
-            extraData_3: string,
-            txData: TxData = {},
-        ): Promise<number> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-            );
-            const gas = await promisify<number>(
-                self.web3ContractInstance.receiveApproval.estimateGas, self.web3ContractInstance,
-            )(
-                from_0,
-                value_1,
-                token_2,
-                extraData_3,
-                txDataWithDefaults,
-            );
-            return gas;
-        },
-        getABIEncodedTransactionData(
-            from_0: string,
-            value_1: BigNumber,
-            token_2: string,
-            extraData_3: string,
-        ): string {
-            const self = this as AuthenticatedProxyContract;
-            const abiEncodedTransactionData = self.web3ContractInstance.receiveApproval.getData(
-                from_0,
-                value_1,
-                token_2,
-                extraData_3,
-            );
-            return abiEncodedTransactionData;
-        },
-    };
-    public initialize = {
-        async sendTransactionAsync(
-            addrUser_0: string,
-            addrRegistry_1: string,
-            txData: TxData = {},
-        ): Promise<string> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-                self.initialize.estimateGasAsync.bind(
-                    self,
-                    addrUser_0,
-                    addrRegistry_1,
-                ),
-            );
-            const txHash = await promisify<string>(
-                self.web3ContractInstance.initialize, self.web3ContractInstance,
-            )(
-                addrUser_0,
-                addrRegistry_1,
-                txDataWithDefaults,
-            );
-            return txHash;
-        },
-        async estimateGasAsync(
-            addrUser_0: string,
-            addrRegistry_1: string,
-            txData: TxData = {},
-        ): Promise<number> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-            );
-            const gas = await promisify<number>(
-                self.web3ContractInstance.initialize.estimateGas, self.web3ContractInstance,
-            )(
-                addrUser_0,
-                addrRegistry_1,
-                txDataWithDefaults,
-            );
-            return gas;
-        },
-        getABIEncodedTransactionData(
-            addrUser_0: string,
-            addrRegistry_1: string,
-        ): string {
-            const self = this as AuthenticatedProxyContract;
-            const abiEncodedTransactionData = self.web3ContractInstance.initialize.getData(
-                addrUser_0,
-                addrRegistry_1,
-            );
-            return abiEncodedTransactionData;
-        },
-    };
-    public setRevoke = {
-        async sendTransactionAsync(
-            revoke_0: boolean,
-            txData: TxData = {},
-        ): Promise<string> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-                self.setRevoke.estimateGasAsync.bind(
-                    self,
-                    revoke_0,
-                ),
-            );
-            const txHash = await promisify<string>(
-                self.web3ContractInstance.setRevoke, self.web3ContractInstance,
-            )(
-                revoke_0,
-                txDataWithDefaults,
-            );
-            return txHash;
-        },
-        async estimateGasAsync(
-            revoke_0: boolean,
-            txData: TxData = {},
-        ): Promise<number> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-            );
-            const gas = await promisify<number>(
-                self.web3ContractInstance.setRevoke.estimateGas, self.web3ContractInstance,
-            )(
-                revoke_0,
-                txDataWithDefaults,
-            );
-            return gas;
-        },
-        getABIEncodedTransactionData(
-            revoke_0: boolean,
-        ): string {
-            const self = this as AuthenticatedProxyContract;
-            const abiEncodedTransactionData = self.web3ContractInstance.setRevoke.getData(
-                revoke_0,
-            );
-            return abiEncodedTransactionData;
-        },
-    };
-    public proxy = {
-        async sendTransactionAsync(
-            dest_0: string,
-            howToCall_1: number|BigNumber,
-            calldata_2: string,
-            txData: TxData = {},
-        ): Promise<string> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-                self.proxy.estimateGasAsync.bind(
-                    self,
-                    dest_0,
-                    howToCall_1,
-                    calldata_2,
-                ),
-            );
-            const txHash = await promisify<string>(
-                self.web3ContractInstance.proxy, self.web3ContractInstance,
-            )(
-                dest_0,
-                howToCall_1,
-                calldata_2,
-                txDataWithDefaults,
-            );
-            return txHash;
-        },
-        async estimateGasAsync(
-            dest_0: string,
-            howToCall_1: number|BigNumber,
-            calldata_2: string,
-            txData: TxData = {},
-        ): Promise<number> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-            );
-            const gas = await promisify<number>(
-                self.web3ContractInstance.proxy.estimateGas, self.web3ContractInstance,
-            )(
-                dest_0,
-                howToCall_1,
-                calldata_2,
-                txDataWithDefaults,
-            );
-            return gas;
-        },
-        getABIEncodedTransactionData(
-            dest_0: string,
-            howToCall_1: number|BigNumber,
-            calldata_2: string,
-        ): string {
-            const self = this as AuthenticatedProxyContract;
-            const abiEncodedTransactionData = self.web3ContractInstance.proxy.getData(
-                dest_0,
-                howToCall_1,
-                calldata_2,
-            );
-            return abiEncodedTransactionData;
-        },
-    };
-    public proxyAssert = {
-        async sendTransactionAsync(
-            dest_0: string,
-            howToCall_1: number|BigNumber,
-            calldata_2: string,
-            txData: TxData = {},
-        ): Promise<string> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-                self.proxyAssert.estimateGasAsync.bind(
-                    self,
-                    dest_0,
-                    howToCall_1,
-                    calldata_2,
-                ),
-            );
-            const txHash = await promisify<string>(
-                self.web3ContractInstance.proxyAssert, self.web3ContractInstance,
-            )(
-                dest_0,
-                howToCall_1,
-                calldata_2,
-                txDataWithDefaults,
-            );
-            return txHash;
-        },
-        async estimateGasAsync(
-            dest_0: string,
-            howToCall_1: number|BigNumber,
-            calldata_2: string,
-            txData: TxData = {},
-        ): Promise<number> {
-            const self = this as AuthenticatedProxyContract;
-            const txDataWithDefaults = await self.applyDefaultsToTxDataAsync(
-                txData,
-            );
-            const gas = await promisify<number>(
-                self.web3ContractInstance.proxyAssert.estimateGas, self.web3ContractInstance,
-            )(
-                dest_0,
-                howToCall_1,
-                calldata_2,
-                txDataWithDefaults,
-            );
-            return gas;
-        },
-        getABIEncodedTransactionData(
-            dest_0: string,
-            howToCall_1: number|BigNumber,
-            calldata_2: string,
-        ): string {
-            const self = this as AuthenticatedProxyContract;
-            const abiEncodedTransactionData = self.web3ContractInstance.proxyAssert.getData(
-                dest_0,
-                howToCall_1,
-                calldata_2,
-            );
-            return abiEncodedTransactionData;
-        },
-    };
-    constructor(web3ContractInstance: Web3.ContractInstance, defaults: Partial<TxData>) {
-        super(web3ContractInstance, defaults);
-        classUtils.bindAll(this, ['web3ContractInstance', 'defaults']);
+    /**
+     * @ignore
+     */
+public static deployedBytecode: string | undefined;
+public static contractName = 'AuthenticatedProxy';
+    private readonly _methodABIIndex: { [name: string]: number } = {};
+private readonly _subscriptionManager: SubscriptionManager<AuthenticatedProxyEventArgs, AuthenticatedProxyEvents>;
+public static async deployFrom0xArtifactAsync(
+        artifact: ContractArtifact | SimpleContractArtifact,
+        supportedProvider: SupportedProvider,
+        txDefaults: Partial<TxData>,
+        logDecodeDependencies: { [contractName: string]: (ContractArtifact | SimpleContractArtifact) },
+    ): Promise<AuthenticatedProxyContract> {
+        assert.doesConformToSchema('txDefaults', txDefaults, schemas.txDataSchema);
+        if (artifact.compilerOutput === undefined) {
+            throw new Error('Compiler output not found in the artifact file');
+        }
+        const provider = providerUtils.standardizeOrThrow(supportedProvider);
+        const bytecode = artifact.compilerOutput.evm.bytecode.object;
+        const abi = artifact.compilerOutput.abi;
+        const logDecodeDependenciesAbiOnly: { [contractName: string]: ContractAbi } = {};
+        if (Object.keys(logDecodeDependencies) !== undefined) {
+            for (const key of Object.keys(logDecodeDependencies)) {
+                logDecodeDependenciesAbiOnly[key] = logDecodeDependencies[key].compilerOutput.abi;
+            }
+        }
+        return AuthenticatedProxyContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly, );
     }
-} // tslint:disable:max-file-line-count
+
+    public static async deployWithLibrariesFrom0xArtifactAsync(
+        artifact: ContractArtifact,
+        libraryArtifacts: { [libraryName: string]: ContractArtifact },
+        supportedProvider: SupportedProvider,
+        txDefaults: Partial<TxData>,
+        logDecodeDependencies: { [contractName: string]: (ContractArtifact | SimpleContractArtifact) },
+    ): Promise<AuthenticatedProxyContract> {
+        assert.doesConformToSchema('txDefaults', txDefaults, schemas.txDataSchema);
+        if (artifact.compilerOutput === undefined) {
+            throw new Error('Compiler output not found in the artifact file');
+        }
+        const provider = providerUtils.standardizeOrThrow(supportedProvider);
+        const abi = artifact.compilerOutput.abi;
+        const logDecodeDependenciesAbiOnly: { [contractName: string]: ContractAbi } = {};
+        if (Object.keys(logDecodeDependencies) !== undefined) {
+            for (const key of Object.keys(logDecodeDependencies)) {
+                logDecodeDependenciesAbiOnly[key] = logDecodeDependencies[key].compilerOutput.abi;
+            }
+        }
+        const libraryAddresses = await AuthenticatedProxyContract._deployLibrariesAsync(
+            artifact,
+            libraryArtifacts,
+            new Web3Wrapper(provider),
+            txDefaults
+        );
+        const bytecode = linkLibrariesInBytecode(
+            artifact,
+            libraryAddresses,
+        );
+        return AuthenticatedProxyContract.deployAsync(bytecode, abi, provider, txDefaults, logDecodeDependenciesAbiOnly, );
+    }
+
+    public static async deployAsync(
+        bytecode: string,
+        abi: ContractAbi,
+        supportedProvider: SupportedProvider,
+        txDefaults: Partial<TxData>,
+        logDecodeDependencies: { [contractName: string]: ContractAbi },
+    ): Promise<AuthenticatedProxyContract> {
+        assert.isHexString('bytecode', bytecode);
+        assert.doesConformToSchema('txDefaults', txDefaults, schemas.txDataSchema);
+        const provider = providerUtils.standardizeOrThrow(supportedProvider);
+        const constructorAbi = BaseContract._lookupConstructorAbi(abi);
+        [] = BaseContract._formatABIDataItemList(
+            constructorAbi.inputs,
+            [],
+            BaseContract._bigNumberToString,
+        );
+        const iface = new ethers.utils.Interface(abi);
+        const deployInfo = iface.deployFunction;
+        const txData = deployInfo.encode(bytecode, []);
+        const web3Wrapper = new Web3Wrapper(provider);
+        const txDataWithDefaults = await BaseContract._applyDefaultsToContractTxDataAsync(
+            {
+                data: txData,
+                ...txDefaults,
+            },
+            web3Wrapper.estimateGasAsync.bind(web3Wrapper),
+        );
+        const txHash = await web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+        logUtils.log(`transactionHash: ${txHash}`);
+        const txReceipt = await web3Wrapper.awaitTransactionSuccessAsync(txHash);
+        logUtils.log(`AuthenticatedProxy successfully deployed at ${txReceipt.contractAddress}`);
+        const contractInstance = new AuthenticatedProxyContract(txReceipt.contractAddress as string, provider, txDefaults, logDecodeDependencies);
+        contractInstance.constructorArgs = [];
+        return contractInstance;
+    }
+
+    /**
+     * @returns      The contract ABI
+     */
+    public static ABI(): ContractAbi {
+        const abi = [
+            { 
+                constant: true,
+                inputs: [
+                ],
+                name: 'proxyType',
+                outputs: [
+                    {
+                        name: 'proxyTypeId',
+                        type: 'uint256',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'pure',
+                type: 'function',
+            },
+            { 
+                constant: true,
+                inputs: [
+                ],
+                name: 'user',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            { 
+                constant: true,
+                inputs: [
+                ],
+                name: 'implementation',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            { 
+                constant: true,
+                inputs: [
+                ],
+                name: 'revoked',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'bool',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            { 
+                constant: true,
+                inputs: [
+                ],
+                name: 'upgradeabilityOwner',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            { 
+                constant: true,
+                inputs: [
+                ],
+                name: 'registry',
+                outputs: [
+                    {
+                        name: '',
+                        type: 'address',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'view',
+                type: 'function',
+            },
+            { 
+                constant: false,
+                inputs: [
+                    {
+                        name: 'from',
+                        type: 'address',
+                    },
+                    {
+                        name: 'value',
+                        type: 'uint256',
+                    },
+                    {
+                        name: 'token',
+                        type: 'address',
+                    },
+                    {
+                        name: 'extraData',
+                        type: 'bytes',
+                    },
+                ],
+                name: 'receiveApproval',
+                outputs: [
+                ],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            { 
+                inputs: [
+                ],
+                outputs: [
+                ],
+                payable: true,
+                stateMutability: 'payable',
+                type: 'fallback',
+            },
+            { 
+                anonymous: false,
+                inputs: [
+                    {
+                        name: 'revoked',
+                        type: 'bool',
+                        indexed: false,
+                    },
+                ],
+                name: 'Revoked',
+                outputs: [
+                ],
+                type: 'event',
+            },
+            { 
+                anonymous: false,
+                inputs: [
+                    {
+                        name: 'sender',
+                        type: 'address',
+                        indexed: true,
+                    },
+                    {
+                        name: 'amount',
+                        type: 'uint256',
+                        indexed: false,
+                    },
+                ],
+                name: 'ReceivedEther',
+                outputs: [
+                ],
+                type: 'event',
+            },
+            { 
+                anonymous: false,
+                inputs: [
+                    {
+                        name: 'from',
+                        type: 'address',
+                        indexed: true,
+                    },
+                    {
+                        name: 'value',
+                        type: 'uint256',
+                        indexed: false,
+                    },
+                    {
+                        name: 'token',
+                        type: 'address',
+                        indexed: true,
+                    },
+                    {
+                        name: 'extraData',
+                        type: 'bytes',
+                        indexed: false,
+                    },
+                ],
+                name: 'ReceivedTokens',
+                outputs: [
+                ],
+                type: 'event',
+            },
+            { 
+                constant: false,
+                inputs: [
+                    {
+                        name: 'addrUser',
+                        type: 'address',
+                    },
+                    {
+                        name: 'addrRegistry',
+                        type: 'address',
+                    },
+                ],
+                name: 'initialize',
+                outputs: [
+                ],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            { 
+                constant: false,
+                inputs: [
+                    {
+                        name: 'revoke',
+                        type: 'bool',
+                    },
+                ],
+                name: 'setRevoke',
+                outputs: [
+                ],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            { 
+                constant: false,
+                inputs: [
+                    {
+                        name: 'dest',
+                        type: 'address',
+                    },
+                    {
+                        name: 'howToCall',
+                        type: 'uint8',
+                    },
+                    {
+                        name: 'calldata',
+                        type: 'bytes',
+                    },
+                ],
+                name: 'proxy',
+                outputs: [
+                    {
+                        name: 'result',
+                        type: 'bool',
+                    },
+                ],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+            { 
+                constant: false,
+                inputs: [
+                    {
+                        name: 'dest',
+                        type: 'address',
+                    },
+                    {
+                        name: 'howToCall',
+                        type: 'uint8',
+                    },
+                    {
+                        name: 'calldata',
+                        type: 'bytes',
+                    },
+                ],
+                name: 'proxyAssert',
+                outputs: [
+                ],
+                payable: false,
+                stateMutability: 'nonpayable',
+                type: 'function',
+            },
+        ] as ContractAbi;
+        return abi;
+    }
+
+    protected static async _deployLibrariesAsync(
+        artifact: ContractArtifact,
+        libraryArtifacts: { [libraryName: string]: ContractArtifact },
+        web3Wrapper: Web3Wrapper,
+        txDefaults: Partial<TxData>,
+        libraryAddresses: { [libraryName: string]: string } = {},
+    ): Promise<{ [libraryName: string]: string }> {
+        const links = artifact.compilerOutput.evm.bytecode.linkReferences;
+        // Go through all linked libraries, recursively deploying them if necessary.
+        for (const link of Object.values(links)) {
+            for (const libraryName of Object.keys(link)) {
+                if (!libraryAddresses[libraryName]) {
+                    // Library not yet deployed.
+                    const libraryArtifact = libraryArtifacts[libraryName];
+                    if (!libraryArtifact) {
+                        throw new Error(`Missing artifact for linked library "${libraryName}"`);
+                    }
+                    // Deploy any dependent libraries used by this library.
+                    await AuthenticatedProxyContract._deployLibrariesAsync(
+                        libraryArtifact,
+                        libraryArtifacts,
+                        web3Wrapper,
+                        txDefaults,
+                        libraryAddresses,
+                    );
+                    // Deploy this library.
+                    const linkedLibraryBytecode = linkLibrariesInBytecode(
+                        libraryArtifact,
+                        libraryAddresses,
+                    );
+                    const txDataWithDefaults = await BaseContract._applyDefaultsToContractTxDataAsync(
+                        {
+                            data: linkedLibraryBytecode,
+                            ...txDefaults,
+                        },
+                        web3Wrapper.estimateGasAsync.bind(web3Wrapper),
+                    );
+                    const txHash = await web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+                    logUtils.log(`transactionHash: ${txHash}`);
+                    const { contractAddress } = await web3Wrapper.awaitTransactionSuccessAsync(txHash);
+                    logUtils.log(`${libraryArtifact.contractName} successfully deployed at ${contractAddress}`);
+                    libraryAddresses[libraryArtifact.contractName] = contractAddress as string;
+                }
+            }
+        }
+        return libraryAddresses;
+    }
+
+    public getFunctionSignature(methodName: string): string {
+        const index = this._methodABIIndex[methodName];
+        const methodAbi = AuthenticatedProxyContract.ABI()[index] as MethodAbi; // tslint:disable-line:no-unnecessary-type-assertion
+        const functionSignature = methodAbiToFunctionSignature(methodAbi);
+        return functionSignature;
+    }
+
+    public getABIDecodedTransactionData<T>(methodName: string, callData: string): T {
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = (this as any) as AuthenticatedProxyContract;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        const abiDecodedCallData = abiEncoder.strictDecode<T>(callData);
+        return abiDecodedCallData;
+    }
+
+    public getABIDecodedReturnData<T>(methodName: string, callData: string): T {
+        if (this._encoderOverrides.decodeOutput) {
+            return this._encoderOverrides.decodeOutput(methodName, callData);
+        }
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = (this as any) as AuthenticatedProxyContract;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        const abiDecodedCallData = abiEncoder.strictDecodeReturnValue<T>(callData);
+        return abiDecodedCallData;
+    }
+
+    public getSelector(methodName: string): string {
+        const functionSignature = this.getFunctionSignature(methodName);
+        const self = (this as any) as AuthenticatedProxyContract;
+        const abiEncoder = self._lookupAbiEncoder(functionSignature);
+        return abiEncoder.getSelector();
+    }
+
+    public proxyType(
+    ): ContractFunctionObj<BigNumber
+> {
+        const self = this as any as AuthenticatedProxyContract;
+        const functionSignature = 'proxyType()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<BigNumber
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                let rawCallResult;
+                if (self._deployedBytecodeIfExists) {
+                    rawCallResult = await self._evmExecAsync(this.getABIEncodedTransactionData());
+                } else {
+                    rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                }
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<BigNumber
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    public user(
+    ): ContractFunctionObj<string
+> {
+        const self = this as any as AuthenticatedProxyContract;
+        const functionSignature = 'user()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<string
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<string
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    public implementation(
+    ): ContractFunctionObj<string
+> {
+        const self = this as any as AuthenticatedProxyContract;
+        const functionSignature = 'implementation()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<string
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<string
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    public revoked(
+    ): ContractFunctionObj<boolean
+> {
+        const self = this as any as AuthenticatedProxyContract;
+        const functionSignature = 'revoked()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<boolean
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<boolean
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    public upgradeabilityOwner(
+    ): ContractFunctionObj<string
+> {
+        const self = this as any as AuthenticatedProxyContract;
+        const functionSignature = 'upgradeabilityOwner()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<string
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<string
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    public registry(
+    ): ContractFunctionObj<string
+> {
+        const self = this as any as AuthenticatedProxyContract;
+        const functionSignature = 'registry()';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<string
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<string
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, []);
+            },
+        };
+    }
+    public receiveApproval(
+            from: string,
+            value: BigNumber,
+            token: string,
+            extraData: string,
+    ): ContractTxFunctionObj<void
+> {
+        const self = this as any as AuthenticatedProxyContract;
+            assert.isString('from', from);
+            assert.isBigNumber('value', value);
+            assert.isString('token', token);
+            assert.isString('extraData', extraData);
+        const functionSignature = 'receiveApproval(address,uint256,address,bytes)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async sendTransactionAsync(
+                txData?: Partial<TxData> | undefined,
+                opts: SendTransactionOpts = { shouldValidate: true },
+            ): Promise<string> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData },
+                    this.estimateGasAsync.bind(this),
+                );
+                if (opts.shouldValidate !== false) {
+                    await this.callAsync(txDataWithDefaults);
+                }
+                return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+            },
+            awaitTransactionSuccessAsync(
+                txData?: Partial<TxData>,
+                opts: AwaitTransactionSuccessOpts = { shouldValidate: true },
+            ): PromiseWithTransactionHash<TransactionReceiptWithDecodedLogs> {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            async estimateGasAsync(
+                txData?: Partial<TxData> | undefined,
+            ): Promise<number> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+            },
+            async createAccessListAsync(
+                txData?: Partial<TxData> | undefined,
+                defaultBlock?: BlockParam,
+            ): Promise<TxAccessListWithGas> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+            },
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<void
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<void
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, [from.toLowerCase(),
+            value,
+            token.toLowerCase(),
+            extraData
+            ]);
+            },
+        };
+    }
+    public initialize(
+            addrUser: string,
+            addrRegistry: string,
+    ): ContractTxFunctionObj<void
+> {
+        const self = this as any as AuthenticatedProxyContract;
+            assert.isString('addrUser', addrUser);
+            assert.isString('addrRegistry', addrRegistry);
+        const functionSignature = 'initialize(address,address)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async sendTransactionAsync(
+                txData?: Partial<TxData> | undefined,
+                opts: SendTransactionOpts = { shouldValidate: true },
+            ): Promise<string> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData },
+                    this.estimateGasAsync.bind(this),
+                );
+                if (opts.shouldValidate !== false) {
+                    await this.callAsync(txDataWithDefaults);
+                }
+                return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+            },
+            awaitTransactionSuccessAsync(
+                txData?: Partial<TxData>,
+                opts: AwaitTransactionSuccessOpts = { shouldValidate: true },
+            ): PromiseWithTransactionHash<TransactionReceiptWithDecodedLogs> {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            async estimateGasAsync(
+                txData?: Partial<TxData> | undefined,
+            ): Promise<number> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+            },
+            async createAccessListAsync(
+                txData?: Partial<TxData> | undefined,
+                defaultBlock?: BlockParam,
+            ): Promise<TxAccessListWithGas> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+            },
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<void
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<void
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, [addrUser.toLowerCase(),
+            addrRegistry.toLowerCase()
+            ]);
+            },
+        };
+    }
+    public setRevoke(
+            revoke: boolean,
+    ): ContractTxFunctionObj<void
+> {
+        const self = this as any as AuthenticatedProxyContract;
+            assert.isBoolean('revoke', revoke);
+        const functionSignature = 'setRevoke(bool)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async sendTransactionAsync(
+                txData?: Partial<TxData> | undefined,
+                opts: SendTransactionOpts = { shouldValidate: true },
+            ): Promise<string> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData },
+                    this.estimateGasAsync.bind(this),
+                );
+                if (opts.shouldValidate !== false) {
+                    await this.callAsync(txDataWithDefaults);
+                }
+                return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+            },
+            awaitTransactionSuccessAsync(
+                txData?: Partial<TxData>,
+                opts: AwaitTransactionSuccessOpts = { shouldValidate: true },
+            ): PromiseWithTransactionHash<TransactionReceiptWithDecodedLogs> {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            async estimateGasAsync(
+                txData?: Partial<TxData> | undefined,
+            ): Promise<number> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+            },
+            async createAccessListAsync(
+                txData?: Partial<TxData> | undefined,
+                defaultBlock?: BlockParam,
+            ): Promise<TxAccessListWithGas> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+            },
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<void
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<void
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, [revoke
+            ]);
+            },
+        };
+    }
+    public proxy(
+            dest: string,
+            howToCall: number|BigNumber,
+            calldata: string,
+    ): ContractTxFunctionObj<boolean
+> {
+        const self = this as any as AuthenticatedProxyContract;
+            assert.isString('dest', dest);
+            assert.isNumberOrBigNumber('howToCall', howToCall);
+            assert.isString('calldata', calldata);
+        const functionSignature = 'proxy(address,uint8,bytes)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async sendTransactionAsync(
+                txData?: Partial<TxData> | undefined,
+                opts: SendTransactionOpts = { shouldValidate: true },
+            ): Promise<string> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData },
+                    this.estimateGasAsync.bind(this),
+                );
+                if (opts.shouldValidate !== false) {
+                    await this.callAsync(txDataWithDefaults);
+                }
+                return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+            },
+            awaitTransactionSuccessAsync(
+                txData?: Partial<TxData>,
+                opts: AwaitTransactionSuccessOpts = { shouldValidate: true },
+            ): PromiseWithTransactionHash<TransactionReceiptWithDecodedLogs> {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            async estimateGasAsync(
+                txData?: Partial<TxData> | undefined,
+            ): Promise<number> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+            },
+            async createAccessListAsync(
+                txData?: Partial<TxData> | undefined,
+                defaultBlock?: BlockParam,
+            ): Promise<TxAccessListWithGas> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+            },
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<boolean
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<boolean
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, [dest.toLowerCase(),
+            howToCall,
+            calldata
+            ]);
+            },
+        };
+    }
+    public proxyAssert(
+            dest: string,
+            howToCall: number|BigNumber,
+            calldata: string,
+    ): ContractTxFunctionObj<void
+> {
+        const self = this as any as AuthenticatedProxyContract;
+            assert.isString('dest', dest);
+            assert.isNumberOrBigNumber('howToCall', howToCall);
+            assert.isString('calldata', calldata);
+        const functionSignature = 'proxyAssert(address,uint8,bytes)';
+        const selector = self._lookupAbiEncoder(functionSignature).getSelector();
+
+        return {
+            selector,
+            async sendTransactionAsync(
+                txData?: Partial<TxData> | undefined,
+                opts: SendTransactionOpts = { shouldValidate: true },
+            ): Promise<string> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData },
+                    this.estimateGasAsync.bind(this),
+                );
+                if (opts.shouldValidate !== false) {
+                    await this.callAsync(txDataWithDefaults);
+                }
+                return self._web3Wrapper.sendTransactionAsync(txDataWithDefaults);
+            },
+            awaitTransactionSuccessAsync(
+                txData?: Partial<TxData>,
+                opts: AwaitTransactionSuccessOpts = { shouldValidate: true },
+            ): PromiseWithTransactionHash<TransactionReceiptWithDecodedLogs> {
+                return self._promiseWithTransactionHash(this.sendTransactionAsync(txData, opts), opts);
+            },
+            async estimateGasAsync(
+                txData?: Partial<TxData> | undefined,
+            ): Promise<number> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.estimateGasAsync(txDataWithDefaults);
+            },
+            async createAccessListAsync(
+                txData?: Partial<TxData> | undefined,
+                defaultBlock?: BlockParam,
+            ): Promise<TxAccessListWithGas> {
+                const txDataWithDefaults = await self._applyDefaultsToTxDataAsync(
+                    { data: this.getABIEncodedTransactionData(), ...txData }
+                );
+                return self._web3Wrapper.createAccessListAsync(txDataWithDefaults, defaultBlock);
+            },
+            async callAsync(
+                callData: Partial<CallData> = {},
+                defaultBlock?: BlockParam,
+            ): Promise<void
+            > {
+                BaseContract._assertCallParams(callData, defaultBlock);
+                const rawCallResult = await self._performCallAsync({ data: this.getABIEncodedTransactionData(), ...callData }, defaultBlock);
+                const abiEncoder = self._lookupAbiEncoder(functionSignature);
+                BaseContract._throwIfUnexpectedEmptyCallResult(rawCallResult, abiEncoder);
+                return abiEncoder.strictDecodeReturnValue<void
+            >(rawCallResult);
+            },
+            getABIEncodedTransactionData(): string {
+                return self._strictEncodeArguments(functionSignature, [dest.toLowerCase(),
+            howToCall,
+            calldata
+            ]);
+            },
+        };
+    }
+
+    /**
+     * Subscribe to an event type emitted by the AuthenticatedProxy contract.
+     * @param eventName The AuthenticatedProxy contract event you would like to subscribe to.
+     * @param indexFilterValues An object where the keys are indexed args returned by the event and
+     * the value is the value you are interested in. E.g `{maker: aUserAddressHex}`
+     * @param callback Callback that gets called when a log is added/removed
+     * @param isVerbose Enable verbose subscription warnings (e.g recoverable network issues encountered)
+     * @return Subscription token used later to unsubscribe
+     */
+    public subscribe<ArgsType extends AuthenticatedProxyEventArgs>(
+        eventName: AuthenticatedProxyEvents,
+        indexFilterValues: IndexedFilterValues,
+        callback: EventCallback<ArgsType>,
+        isVerbose: boolean = false,
+        blockPollingIntervalMs?: number,
+    ): string {
+        assert.doesBelongToStringEnum('eventName', eventName, AuthenticatedProxyEvents);
+        assert.doesConformToSchema('indexFilterValues', indexFilterValues, schemas.indexFilterValuesSchema);
+        assert.isFunction('callback', callback);
+        const subscriptionToken = this._subscriptionManager.subscribe<ArgsType>(
+            this.address,
+            eventName,
+            indexFilterValues,
+            AuthenticatedProxyContract.ABI(),
+            callback,
+            isVerbose,
+            blockPollingIntervalMs,
+        );
+        return subscriptionToken;
+    }
+
+    /**
+     * Cancel a subscription
+     * @param subscriptionToken Subscription token returned by `subscribe()`
+     */
+    public unsubscribe(subscriptionToken: string): void {
+        this._subscriptionManager.unsubscribe(subscriptionToken);
+    }
+
+    /**
+     * Cancels all existing subscriptions
+     */
+    public unsubscribeAll(): void {
+        this._subscriptionManager.unsubscribeAll();
+    }
+
+    /**
+     * Gets historical logs without creating a subscription
+     * @param eventName The AuthenticatedProxy contract event you would like to subscribe to.
+     * @param blockRange Block range to get logs from.
+     * @param indexFilterValues An object where the keys are indexed args returned by the event and
+     * the value is the value you are interested in. E.g `{_from: aUserAddressHex}`
+     * @return Array of logs that match the parameters
+     */
+    public async getLogsAsync<ArgsType extends AuthenticatedProxyEventArgs>(
+        eventName: AuthenticatedProxyEvents,
+        blockRange: BlockRange,
+        indexFilterValues: IndexedFilterValues,
+    ): Promise<Array<LogWithDecodedArgs<ArgsType>>> {
+        assert.doesBelongToStringEnum('eventName', eventName, AuthenticatedProxyEvents);
+        assert.doesConformToSchema('blockRange', blockRange, schemas.blockRangeSchema);
+        assert.doesConformToSchema('indexFilterValues', indexFilterValues, schemas.indexFilterValuesSchema);
+        const logs = await this._subscriptionManager.getLogsAsync<ArgsType>(
+            this.address,
+            eventName,
+            blockRange,
+            indexFilterValues,
+            AuthenticatedProxyContract.ABI(),
+        );
+        return logs;
+    }
+
+    constructor(
+        address: string,
+        supportedProvider: SupportedProvider,
+        txDefaults?: Partial<TxData>,
+        logDecodeDependencies?: { [contractName: string]: ContractAbi },
+        deployedBytecode: string | undefined = AuthenticatedProxyContract.deployedBytecode,
+        encoderOverrides?: Partial<EncoderOverrides>,
+    ) {
+        super('AuthenticatedProxy', AuthenticatedProxyContract.ABI(), address, supportedProvider, txDefaults, logDecodeDependencies, deployedBytecode, encoderOverrides);
+        classUtils.bindAll(this, ['_abiEncoderByFunctionSignature', 'address', '_web3Wrapper']);
+this._subscriptionManager = new SubscriptionManager<AuthenticatedProxyEventArgs, AuthenticatedProxyEvents>(
+            AuthenticatedProxyContract.ABI(),
+            this._web3Wrapper,
+        );
+AuthenticatedProxyContract.ABI().forEach((item, index) => {
+            if (item.type === 'function') {
+                const methodAbi = item as MethodAbi;
+                this._methodABIIndex[methodAbi.name] = index;
+            }
+        });
+    }
+}
+
+// tslint:disable:max-file-line-count
+// tslint:enable:no-unbound-method no-parameter-reassignment no-consecutive-blank-lines ordered-imports align
+// tslint:enable:trailing-comma whitespace no-trailing-whitespace
