@@ -1,15 +1,16 @@
 /* Sourced from 0x.js */
 
-import { BigNumber } from '@0xproject/utils';
+import { BigNumber } from '@0x/utils';
+import { AbiDefinition, EventAbi, EventParameter, LogEntry } from 'ethereum-types';
 import * as _ from 'lodash';
-import * as Web3 from 'web3';
+import Web3 from 'web3';
 import * as SolidityCoder from 'web3/lib/solidity/coder';
 
 import { AbiType, ContractEventArgs, DecodedLogArgs, LogWithDecodedArgs, RawLog, SolidityTypes } from '../types';
 
 export class AbiDecoder {
-    private _savedABIs: Web3.AbiDefinition[] = [];
-    private _methodIds: { [signatureHash: string]: Web3.EventAbi } = {};
+    private _savedABIs: AbiDefinition[] = [];
+    private _methodIds: { [signatureHash: string]: EventAbi } = {};
     private static _padZeros(address: string) {
         let formatted = address;
         if (_.startsWith(formatted, '0x')) {
@@ -19,12 +20,12 @@ export class AbiDecoder {
         formatted = _.padStart(formatted, 40, '0');
         return `0x${formatted}`;
     }
-    constructor(abiArrays: Web3.AbiDefinition[][]) {
+    constructor(abiArrays: AbiDefinition[][]) {
         _.map(abiArrays, this._addABI.bind(this));
     }
     // This method can only decode logs from the 0x & ERC20 smart contracts
     public tryToDecodeLogOrNoop<ArgsType extends ContractEventArgs>(
-        log: Web3.LogEntry,
+        log: LogEntry,
     ): LogWithDecodedArgs<ArgsType> | RawLog {
         const methodId = log.topics[0];
         const event = this._methodIds[methodId];
@@ -40,7 +41,7 @@ export class AbiDecoder {
         const dataTypes = _.map(nonIndexedInputs, input => input.type);
         const decodedData = SolidityCoder.decodeParams(dataTypes, logData.slice('0x'.length));
 
-        _.map(event.inputs, (param: Web3.EventParameter) => {
+        _.map(event.inputs, (param: EventParameter) => {
             // Indexed parameters are stored in topics. Non-indexed ones in decodedData
             let value = param.indexed ? log.topics[topicsIndex++] : decodedData[dataIndex++];
             if (param.type === SolidityTypes.Address) {
@@ -58,15 +59,17 @@ export class AbiDecoder {
         return {
             ...log,
             event: event.name,
+            // @ts-expect-error TODO: Fix if needed, but insignificant
             args: decodedParams,
         };
     }
-    private _addABI(abiArray: Web3.AbiDefinition[]): void {
-        _.map(abiArray, (abi: Web3.AbiDefinition) => {
+    private _addABI(abiArray: AbiDefinition[]): void {
+        _.map(abiArray, (abi: AbiDefinition) => {
             if (abi.type === AbiType.Event) {
-                const signature = `${abi.name}(${_.map(abi.inputs, input => input.type).join(',')})`;
-                const signatureHash = new Web3().sha3(signature);
-                this._methodIds[signatureHash] = abi;
+                const eventAbi = abi as EventAbi;
+                const signature = `${eventAbi.name}(${_.map(eventAbi.inputs, input => input.type).join(',')})`;
+                const signatureHash = new Web3().utils.sha3(signature) as string;
+                this._methodIds[signatureHash] = eventAbi;
             }
         });
         this._savedABIs = this._savedABIs.concat(abiArray);
